+++ CMakeLists.txt
@@ -69,6 +69,7 @@ include(GNUInstallDirs)
 include_directories(${PROJECT_BINARY_DIR})
 
 add_subdirectory(src)
+add_subdirectory(src/tests)
 
 configure_file("${CMAKE_CURRENT_SOURCE_DIR}/cmake_uninstall.cmake.in"
         "${CMAKE_CURRENT_BINARY_DIR}/cmake_uninstall.cmake" IMMEDIATE @ONLY)
