+++ src/CMakeLists.txt
@@ -52,31 +52,6 @@ set(Scientific_C_H
                 datasets.h
                 modelvalidation.h)
 
-set(NumericTest_SRC testnumeric.c)
-set(StrVectorTest_SRC teststrvector.c)
-set(DVectorTest_SRC testdvector.c)
-set(UIVectorTest_SRC testuivector.c)
-set(IVectorTest_SRC testivector.c)
-set(VectorListTest_SRC testlist.c)
-set(MatrixTest_SRC testmatrix.c)
-set(TensorTest_SRC testtensor.c)
-set(AlgebraTest_SRC testalgebra.c)
-set(PCATest_SRC testpca.c)
-set(CPCATest_SRC testcpca.c)
-set(PLSTest_SRC testpls.c)
-set(ICATest_SRC testica.c)
-set(EPLSTest_SRC testepls.c)
-set(MPCATest_SRC testmwpca.c)
-set(MPLSTest_SRC testmwpls.c)
-set(CLUSTERINGTest_SRC testclustering.c)
-set(MLRTest_SRC testmlr.c)
-set(SVDTest_SRC testsvd.c)
-set(CLASSIFIERTest_SRC testlda.c)
-set(METRICSPACETest_SRC testmetrics.c)
-set(INTERPOLATETest_SRC testinterpolate.c)
-set(GRAPHSTest_SRC testgraphs.c)
-set(OptimizationTest_SRC testoptimization.c)
-
 if(WIN32)
   message(STATUS "Set mingw options")
   set_property(DIRECTORY APPEND PROPERTY COMPILE_DEFINITIONS_RELEASE)
@@ -99,91 +74,3 @@ else()
   install(FILES ${Scientific_C_H} DESTINATION include/scientific)
   install(FILES scientific.h DESTINATION include)
 endif()
-
-add_executable(testnumeric ${NumericTest_SRC})
-target_link_libraries(testnumeric libscientific)
-add_test(NAME testnumeric COMMAND testnumeric)
-
-add_executable(teststrvector ${StrVectorTest_SRC})
-target_link_libraries(teststrvector libscientific)
-add_test(NAME teststrvector COMMAND teststrvector)
-
-add_executable(testdvector ${DVectorTest_SRC})
-target_link_libraries(testdvector libscientific)
-add_test(NAME testdvector COMMAND testdvector)
-
-add_executable(testuivector ${UIVectorTest_SRC})
-target_link_libraries(testuivector libscientific)
-add_test(NAME testuivector COMMAND testuivector)
-
-add_executable(testivector ${IVectorTest_SRC})
-target_link_libraries(testivector libscientific)
-add_test(NAME testivector COMMAND testivector)
-
-add_executable(testlist ${VectorListTest_SRC})
-target_link_libraries(testlist libscientific)
-add_test(NAME testlist COMMAND testlist)
-
-add_executable(testmatrix ${MatrixTest_SRC})
-target_link_libraries(testmatrix libscientific)
-add_test(NAME testmatrix COMMAND testmatrix)
-
-add_executable(testtensor ${TensorTest_SRC})
-target_link_libraries(testtensor libscientific)
-add_test(NAME testtensor COMMAND testtensor)
-
-add_executable(testalgebra ${AlgebraTest_SRC})
-target_link_libraries(testalgebra libscientific)
-add_test(NAME testalgebra COMMAND testalgebra)
-
-add_executable(testpca ${PCATest_SRC})
-target_link_libraries(testpca libscientific)
-add_test(NAME testpca COMMAND testpca)
-
-add_executable(testcpca ${CPCATest_SRC})
-target_link_libraries(testcpca libscientific)
-add_test(NAME testcpca COMMAND testcpca)
-
-add_executable(testpls ${PLSTest_SRC})
[…]
