### 2023-03-22 8de79e6
Remove garbage functions on upls and fixing Division by zero bugs
