+++ src/upls.c
@@ -203,7 +203,7 @@ void UPLS(tensor* X_,
           ssignal *s)
 {
 
-  if(CheckTensors(X_, Y_) == 0){
+  if(CheckTensors(X_, Y_) == 0 && npc >= 1){
     size_t pc, i, j, k, order, column;
     double a, ssx;
     tensor *X;
@@ -2087,680 +2087,6 @@ void UPLSCostPopulation(tensor *ax, tensor *ay,
     DelTensor(&sdep);
   }
   DelTensor(&subax);
-}
-
-void UPSLPSOVariableSelection(tensor *ax, tensor *ay, tensor *px, tensor *py,
-                       size_t xautoscaling, size_t yautoscaling, size_t npc, int validation_type, size_t ngroup, size_t niter, size_t population_size, double randomness,
-                       uivector **varselected, matrix **map, uivector **vardistribution, ssignal *s)
-{
-  size_t i, j, iter, new_position, population_size_, tmp_var_id, nvarON;
-  double model_r2, cost_g_best, cost_best, tmp_model_r2, tmp_cost_best, tmp_g_best, new_v, new_pos, rand1, rand2;
-  matrix *models, *position, *velocity;
-  uivector *popvector, *best_model;
-  dvector *b_position, *g_position, *rowmap;
-
-  /*1) Initialize the population*/
-
-  population_size_ = population_size + 1; /* + 1 is the model with all variables */
-
-  NewMatrix(&models, population_size_, ax->m[0]->col);
-  NewMatrix(&position, population_size_, ax->m[0]->col); /*position chagned stored if are -1 then the position is not stored.... else is  a number >= 0 and < mx->col*/
-  NewMatrix(&velocity, population_size_, ax->m[0]->col);
-  NewDVector(&b_position, ax->m[0]->col);
-  NewDVector(&g_position, ax->m[0]->col);
-  NewUIVector(&popvector, ax->m[0]->col);
-  NewUIVector(&best_model, ax->m[0]->col);
-  NewDVector(&rowmap, 4);
-
-  UIVectorSet(popvector, 1); /*get all variables*/
-
-  UPLSCostPopulation(ax, ay, px, py, popvector, xautoscaling, yautoscaling, npc, validation_type, ngroup, niter, &model_r2, &cost_g_best, s);
-
-  setDVectorValue(rowmap, 0, model_r2);
-  setDVectorValue(rowmap, 1, cost_g_best);
-  setDVectorValue(rowmap, 2, (cost_g_best  * (population_size - ax->m[0]->col -1)) /  (ax->m[0]->col*(1-cost_g_best )));
-  setDVectorValue(rowmap, 3, ax->m[0]->col);
-
-  MatrixAppendRow((*map), rowmap);
-
-  for(i = 0; i < ax->m[0]->col; i++){
-    UIVectorAppend((*vardistribution), 1);
-  }
-
-  /*the best and global position are the same in the first model that have all the variable setted ON*/
-  for(i = 0; i < ax->m[0]->col; i++){
-    /*All variable are global and best...*/
-    setDVectorValue(g_position, i, i);
-    setDVectorValue(b_position, i, i);
-    setMatrixValue(velocity, 0, i, 0.f);
-  }
-
-  /*
-  printf("model 0 with all variable: %f\n", cost_g_best);
-  */
-
-  /*2) Initialize the models particle.*/
-  MatrixSet(models, 1.0); // all variables are setted ON!
-  MatrixSet(position, -1.0); // all variables on model are unvariate...
-  MatrixSet(velocity, 0); // all variables have no speed...
-
-  srand(ax->m[0]->row + ax->m[0]->col + ay->m[0]->col + population_size_);
-
-  for(i = 1; i < population_size_; i++){
-
[…]

[…]
