### 2023-03-22 81b818e
fix bug pca division by zero
