### 2023-03-22 5672075
Fix scientific info sprintf to snprintf. Added a new test. Fixed DModX to be absolute
