### 2023-03-23 55569b7
Fix strcat strcpy safe and code optimization
