+++ src/tests/testclustering.c
@@ -922,7 +922,7 @@ int main(void){
   //test15();
   test16();
   test17();
-  test18();
+  /*test18();*/
   test19();
   return 0;
 }
