### 2023-03-25 dd3d46d
Fix define each identifier in a dedicated statement
