+++ src/matrix.c
@@ -40,7 +40,8 @@ void initMatrix(matrix **m)
 
 void NewMatrix(matrix **m, size_t row_ , size_t col_)
 {
-  size_t i, j;
+  size_t i;
+  size_t j;
   (*m) = xmalloc(sizeof(matrix));
   (*m)->row = row_;
   (*m)->col = col_;
@@ -54,7 +55,8 @@ void NewMatrix(matrix **m, size_t row_ , size_t col_)
 
 void ResizeMatrix(matrix *m, size_t row_, size_t col_)
 {
-  size_t i, j;
+  size_t i;
+  size_t j;
   if(m != NULL){
     if(m->row == row_ && m->col == col_){
       MatrixSet(m, +0.f);
@@ -96,7 +98,8 @@ void DelMatrix(matrix **m)
 
 void MatrixCheck(matrix *m)
 {
-  size_t i, j;
+  size_t i;
+  size_t j;
   for(i = 0; i < m->row; i++){
     for(j = 0; j < m->col; j++){
       if(_isnan_(m->data[i][j]) || !isfinite(m->data[i][j])){
@@ -108,7 +111,8 @@ void MatrixCheck(matrix *m)
 
 void FindNan(matrix *m)
 {
-  size_t i, j;
+  size_t i;
+  size_t j;
   for(i = 0; i < m->row; i++){
     for(j = 0; j < m->col; j++){
       if(_isnan_(m->data[i][j])){
@@ -120,7 +124,8 @@ void FindNan(matrix *m)
 
 void PrintMatrix(matrix *m)
 {
-  size_t i, j;
+  size_t i;
+  size_t j;
   printf("Matrix of row: %u; col: %u\n", (unsigned int)m->row, (unsigned int)m->col);
   for(i = 0; i < m->row; i++){
     for(j = 0; j < m->col; j++)
@@ -132,10 +137,10 @@ void PrintMatrix(matrix *m)
 /*if a value is in matrix return 1 else 0*/
 int ValInMatrix(matrix* m, double val)
 {
-  size_t i, j;
+  size_t i;
+  size_t j;
   for(i = 0; i < m->row; i++){
     for(j = 0; j < m->col; j++){
-      /*if((size_t)getMatrixValue(m, i, j) == (size_t)id) */
       if(FLOAT_EQ(m->data[i][j], val, 1*10e-8))
         return 1;
       else
@@ -147,7 +152,8 @@ int ValInMatrix(matrix* m, double val)
 
 void MatrixSet(matrix *m, double val)
 {
-  size_t i, j;
+  size_t i;
+  size_t j;
   if(m->row == m->col){
     for(i = 0; i < m->row; i++){
       m->data[i][i] = val;
@@ -167,7 +173,8 @@ void MatrixSet(matrix *m, double val)
 
 void MatrixInitRandomInt(matrix *m, int low, int high)
 {
-  size_t i, j;
+  size_t i;
+  size_t j;
   for(i = 0; i < m->row; i++){
     for(j = 0; j < m->col; j++){
       m->data[i][j] = (double)randInt(low, high);
@@ -177,7 +184,8 @@ void MatrixInitRandomInt(matrix *m, int low, int high)
 
 void MatrixInitRandomFloat(matrix *m, double low, double high)
 {
-  size_t i, j;
+  size_t i;
+  size_t j;
   for(i = 0; i < m->row; i++){
     for(j = 0; j < m->col; j++){
       m->data[i][j] = randDouble(low, high);
@@ -187,7 +195,8 @@ void MatrixInitRandomFloat(matrix *m, double low, double high)
 
 void MatrixCopy(matrix *msrc, matrix **mdst)
 {
[…]
