### 2023-03-26 ba26270
Adding more unit tests
