### 2023-03-26 eb12986
Adding more tests for matrix.h
