+++ src/tests/testmatrix.c
@@ -26,18 +26,9 @@
 #include "numeric.h"
 #include "algebra.h"
 /*
-MT_DVectorMatrixDotProduct
-Matrix2ABSMatrix
 Matrix2IntFactorsMatrix
-Matrix2LogMatrix
-Matrix2SQRTMatrix
-Matrix2SquareMatrix
 MatrixAppendUIRow
 MatrixCheck
-MatrixColRMS
-MatrixColSDEV
-MatrixColVar
-MatrixColumnMinMax
 MatrixDeleteColAt
 MatrixDeleteRowAt
 MatrixInitRandomFloat
@@ -47,11 +38,226 @@ MatrixNorm
 MatrixReverseSort
 MatrixRowAverage
 MatrixRowCenterScaling
-MatrixSet
+
 MatrixSort
 RowColOuterProduct
 */
 
+void Test47()
+{
+  puts("Test 47: MatrixSet");
+  matrix *m;
+  size_t i;
+  size_t j;
+  NewMatrix(&m, 3, 3);
+  MatrixSet(m, -1);
+  for(i = 0; i < m->row; i++){
+    for(j = 0; j < m->col; j++){
+      if(!FLOAT_EQ(m->data[i][j], -1.f, 1e-3))
+        abort();
+      else
+        continue;
+    }
+  }
+  puts("MatrixSet: OK");
+  DelMatrix(&m);
+}
+
+void Test46()
+{
+  puts("Test 46: MatrixColumnMinMax");
+  matrix *m;
+  double min;
+  double max;
+  NewMatrix(&m, 3, 3);
+  m->data[0][0] = 1; m->data[0][1] = 0; m->data[0][2] = 3;
+  m->data[1][0] = 11.; m->data[1][1] = 5; m->data[1][2] = 2;
+  m->data[2][0] = 6; m->data[2][1] = 12; m->data[2][2] = 5;
+  
+  MatrixColumnMinMax(m, 0, &min, &max);
+  if(!FLOAT_EQ(min, 1.f, 1e-4) ||
+     !FLOAT_EQ(max, 11.f, 1e-4))
+    abort();
+  MatrixColumnMinMax(m, 1, &min, &max);
+  if(!FLOAT_EQ(min, 0.f, 1e-4) ||
+     !FLOAT_EQ(max, 12.f, 1e-4))
+    abort();
+  MatrixColumnMinMax(m, 1, &min, &max);
+  if(!FLOAT_EQ(min, 0.f, 1e-4) ||
+     !FLOAT_EQ(max, 12.f, 1e-4))
+    abort();
+  
+  puts("MatrixColumnMinMax: OK");
+  DelMatrix(&m);
+}
+
+void Test45()
+{
+  puts("Test 45: MatrixColVar");
+  matrix *m;
+  dvector *v;
+  NewMatrix(&m, 3, 3);
+  m->data[0][0] = 1; m->data[0][1] = 0; m->data[0][2] = 3;
+  m->data[1][0] = 11.; m->data[1][1] = 5; m->data[1][2] = 2;
+  m->data[2][0] = 6; m->data[2][1] = 12; m->data[2][2] = 5;
+  initDVector(&v);
+  MatrixColVar(m, v);
+  if(!FLOAT_EQ(v->data[0], 25.0000, 1e-4) ||
+     !FLOAT_EQ(v->data[1], 36.3333, 1e-4) ||
+     !FLOAT_EQ(v->data[2], 2.3333, 1e-4))
+     abort();
+  puts("MatrixColVar: OK");
+  DelMatrix(&m);
+  DelDVector(&v);
+}
+
+void Test44()
+{
+  puts("Test 44: MatrixColSDEV");
+  matrix *m;
+  dvector *v;
+  NewMatrix(&m, 3, 3);
[…]
