+++ src/tests/testmetrics.c
@@ -1,3 +1,4 @@
+
 /* testmetrics.c
 *
 * Copyright (C) <2016>  Giuseppe Marco Randazzo
@@ -23,8 +24,95 @@
 #include "matrix.h"
 #include "statistic.h"
 
+
+/*
+void EuclideanDistanceCondensed(matrix* m, dvector *distances, size_t nthreads);
+void SquaredEuclideanDistanceCondensed(matrix *m, dvector *distances, size_t nthreads);
+void ManhattanDistanceCondensed(matrix *m, dvector *distances, size_t nthreads);
+*/
+
+void test9()
+{
+  puts("Test9: MatrixMahalanobisDistance");
+  /* Euclidean distance between two matrix test */
+  matrix *m1;
+  matrix *m2;
+  double mxmxdist;
+
+  NewMatrix(&m1, 4, 3);
+  m1->data[0][0] = 1; m1->data[0][1] = 2; m1->data[0][2] = 3;
+  m1->data[1][0] = 4; m1->data[1][1] = 5; m1->data[1][2] = 6;
+  m1->data[2][0] = 7; m1->data[2][1] = 8; m1->data[2][2] = 11;
+  m1->data[3][0] = 9; m1->data[3][1] = 10; m1->data[3][2] = 12;
+
+  NewMatrix(&m2, 6, 3);
+  m2->data[0][0] = 1; m2->data[0][1] = 2; m2->data[0][2] = 3;
+  m2->data[1][0] = 1; m2->data[1][1] = 8; m2->data[1][2] = 1;
+  m2->data[2][0] = 4; m2->data[2][1] = 0; m2->data[2][2] = 7;
+  m2->data[3][0] = 5; m2->data[3][1] = 1; m2->data[3][2] = 6;
+  m2->data[4][0] = 3; m2->data[4][1] = 3; m2->data[4][2] = 4;
+  m2->data[5][0] = 5; m2->data[5][1] = 2; m2->data[5][2] = 7;
+
+  mxmxdist = MatrixMahalanobisDistance(m1, m2);
+  if(!FLOAT_EQ(mxmxdist, 1.931663, 1e-4))
+    abort();
+  puts("MatrixMahalanobisDistance: OK");
+  DelMatrix(&m1);
+  DelMatrix(&m2);
+}
+void test8()
+{
+  puts("Test8: CosineDistance");
+  size_t i, j, maxrow, maxcol;
+  matrix *m, *dist;
+
+  maxrow = 600;
+  maxcol = 10;
+  srand_(maxrow);
+  NewMatrix(&m, maxrow, maxcol);
+
+  srand(maxrow+maxcol);
+  for(i = 0; i < maxrow; i++){
+    for(j = 0; j < maxcol; j++){
+      setMatrixValue(m, i, j, rand() % 100);
+    }
+  }
+
+  initMatrix(&dist);
+  CosineDistance(m, m, dist, 4);
+  puts("CosineDistance: OK");
+  DelMatrix(&m);
+  DelMatrix(&dist);
+}
+
 void test7()
 {
+  puts("Test7: SquaredEuclideanDistance");
+  size_t i, j, maxrow, maxcol;
+  matrix *m, *dist;
+
+  maxrow = 600;
+  maxcol = 10;
+  srand_(maxrow);
+  NewMatrix(&m, maxrow, maxcol);
+
+  srand(maxrow+maxcol);
+  for(i = 0; i < maxrow; i++){
+    for(j = 0; j < maxcol; j++){
+      setMatrixValue(m, i, j, rand() % 100);
+    }
+  }
+
+  initMatrix(&dist);
+  SquaredEuclideanDistance(m, m, dist, 4);
+  puts("SquaredEuclideanDistance: OK");
+  DelMatrix(&m);
[…]
