### 2023-03-28 84c5a63
Adding documentation to tensor and fixing some bloat documentation/code
