+++ src/tensor.h
@@ -35,51 +35,141 @@ typedef struct{
   size_t order;
 }tensor;
 
+/*
+ * Description:
+ * initialize an empty tensor
+ */
 void initTensor(tensor **t);
+
+/*
+ * Description:
+ * Create a new empty tensor with a predefined order.
+ */
 void NewTensor(tensor** t, size_t order_);
+
+/*
+ * Description:
+ * Create into an empty tensor a matrix of "row x col" at position "order"
+ */
 void NewTensorMatrix(tensor *t, size_t order, size_t row, size_t col);
+
+/*
+ * Description:
+ * Append to a tensor a matrix at the end of it.
+ */
 void AddTensorMatrix(tensor *t, size_t row, size_t col);
 
+/*
+ * Description:
+ * Delete a tensor and free up the memory.
+ */
 void DelTensor(tensor**t);
 
+/*
+ * Description:
+ * Print to video the tensor.
+ */
 void PrintTensor(tensor *t);
+
+/*
+ * Description:
+ * Set a value into a tensor 
+ */
 void setTensorValue(tensor *t, size_t order, size_t row, size_t col, double val);
+
+/*
+ * Description:
+ * Get a value from a tensor
+ */
 double getTensorValue(tensor *t, size_t order, size_t row, size_t col);
+
+/*
+ * Description:
+ * Append a matrix to a tensor at the end of it.
+ */
 void TensorAppendMatrix(tensor *tdst, matrix *msrc);
+
+/*
+ * Description:
+ * Append a matrix to a tensor in a specific position of it;
+ */
 void TensorAppendMatrixAt(tensor *tdst, size_t order, matrix *msrc);
+
+/*
+ * Description:
+ * Append a column vector into the matrix of a tensor at position "order"
+ */
 void TensorAppendColumn(tensor *t, size_t order, dvector* column);
+
+/*
+ * Description:
+ * Append a row vector into the matrix of a tensor at position "order"
+ */
 void TensorAppendRow(tensor *t, size_t order, dvector* row);
+
+/*
+ * Description:
+ * Set all the values of a tensor to a specific value "val"
+ */
 void TensorSet(tensor *t, double val);
+
+/*
+ * Description:
+ * Copy a tensor from a source into a destination
+ */
 void TensorCopy(tensor *asrc, tensor **adst);
 
 
-
-/* matrix *getRowPlane(tensor *t, size_t row);
-matrix  *getColumnPlane(tensor *t, size_t col);
-void TensorAppendRowPlane(tensor **t, matrix *row);
-void TensorAppendColPlane(tensor **t, matrix *col);
-*/
-
-/*  Matrix Operations */
+/*
+ * Description:
+ * Calculate the mean centered tensor
+ */
 void MeanCenteredTensor(tensor *t, tensor *tc);
-void TensorColAverage(tensor *t, matrix *colaverage);
-void TensorColSDEV(tensor *t, matrix *colsdev);
-void TensorTranspose(tensor *t1, tensor *t2);
-
-void TransposedTensorDVectorProduct(tensor *t, dvector *v, matrix *p);
-void DvectorTensorDotProduct(tensor *t, dvector *v, matrix *m);
-void TensorMatrixDotProduct(tensor *t, matrix *m, dvector *v);
-void KronekerProductVectorMatrix(dvector *v, matrix *m, tensor *t);
-
 
 /*
-void MatrixDVectorDotProduct(matrix *m, dvector *v, dvector *r);
-void DVectorMatrixDotProduct(matrix *mx, dvector *v, dvector *p);
-void MatrixDotProduct(matrix *m_t, matrix *m, matrix **r);
-void MatrixTranspose(matrix *m, matrix *r);
[…]
