+++ src/interpolate.h
@@ -22,13 +22,25 @@
 #include "matrix.h"
 #include "vector.h"
 
-/* Description: calculate the natural cubic spline interpolation equations */
+/* Description: calculate the natural cubic spline interpolation equations 
+ * 
+ * Algorithm reference:
+ * Natural Cubic Spline: Numerical Analysis book Richard L. Burden and J. Douglas Faires
+ * pag. 149
+ * ISBN-13: 978-0-538-73351-9
+ * 
+ * N.B.: If two points have different y but share same x the algorithm will fail
+ */
 void cubic_spline_interpolation(matrix *xy, matrix *S);
 
-/* Description: predict using the  natural cubic spline interpolation equations a vector of x */
+/* Description: predict using the  natural cubic spline 
+ *              interpolation equations a vector of x
+ */
 void cubic_spline_predict(dvector *x_, matrix *S, dvector *y_pred);
 
-/* Description: interpolate x and y using the natural cubic spline equations and get directly the interpolation. */
+/* Description: interpolate x and y using the natural cubic 
+ *              spline equations and get directly the interpolation.
+ */
 void interpolate(matrix *xy, size_t npoints, matrix *interp_xy);
 
 #endif
