### 2023-04-18 044eda2
Fix minor bug in libscientific matrix generation from a single colum list
