### 2023-04-22 98dca4c
Get beta coefficients from a PLS model
