### 2023-04-24 ee3bea0
Adding clustering method python ffi; code improvement in metricspace; adding more tests metricspace
