+++ src/python_bindings/libscientific/vector.py
@@ -53,6 +53,49 @@ class DVECTOR(ctypes.Structure):
         return self.__class__.__name__
 
 
+class UIVECTOR(ctypes.Structure):
+    """
+    unsigned int vector data structure
+    """
+    _fields_ = [
+        ("data", ctypes.POINTER(ctypes.c_size_t)),
+        ("size", ctypes.c_size_t)]
+
+    def __repr__(self):
+        return self.__class__.__name__
+
+    def __str__(self):
+        return self.__class__.__name__
+
+
+class IVECTOR(ctypes.Structure):
+    """
+    unsigned int vector data structure
+    """
+    _fields_ = [
+        ("data", ctypes.POINTER(ctypes.c_int)),
+        ("size", ctypes.c_size_t)]
+
+    def __repr__(self):
+        return self.__class__.__name__
+
+    def __str__(self):
+        return self.__class__.__name__
+
+
+def xvector_tolist(dvect):
+    """
+    Convert a libscientific double/unsigned int/int vector into a python list
+    """
+    d_lst = []
+    try:
+        for i in range(dvect[0].size):
+            d_lst.append(dvect[0].data[i])
+    except TypeError:
+        for i in range(dvect.size):
+            d_lst.append(dvect.data[i])
+    return d_lst
+
 lsci.initDVector.argtypes = [ctypes.POINTER(ctypes.POINTER(DVECTOR))]
 lsci.initDVector.restype = None
 
@@ -98,7 +141,7 @@ def del_dvector(dvect):
     """
     DelDVector: Delete an allocated libscientific double vector
     """
-    lsci.initDVector(ctypes.pointer(dvect))
+    lsci.DelDVector(ctypes.pointer(dvect))
 
 lsci.DVectorResize.argtypes = [ctypes.POINTER(DVECTOR),
                                ctypes.c_size_t]
@@ -174,20 +217,6 @@ def get_dvector_value(dvect, indx):
     return lsci.getDVectorValue(dvect, indx)
 
 
-def dvector_tolist(dvect):
-    """
-    Convert a libscientific double vector into a python list
-    """
-    d_lst = []
-    try:
-        for i in range(dvect[0].size):
-            d_lst.append(dvect[0].data[i])
-    except TypeError:
-        for i in range(dvect.size):
-            d_lst.append(dvect.data[i])
-    return d_lst
-
-
 lsci.DVectorAppend.argtypes = [ctypes.POINTER(DVECTOR),
                                ctypes.c_double]
 lsci.DVectorAppend.restype = None
@@ -197,6 +226,13 @@ def dvector_append(dvect, val):
     Append a value to a double vector d
     """
     return lsci.DVectorAppend(dvect, val)
+
+
+def dvector_tolist(dvect):
+    """
+    Double vector to python list
+    """
+    return xvector_tolist(dvect)
 
 
 lsci.DVectorRemoveAt.argtypes = [ctypes.POINTER(DVECTOR),
@@ -220,6 +256,245 @@ def dvector_copy(dvect_src):
     dvect_dst = init_dvector()
     lsci.DVectorCopy(dvect_src, dvect_dst)
     return dvect_dst
+
+
+lsci.initUIVector.argtypes = [ctypes.POINTER(ctypes.POINTER(UIVECTOR))]
+lsci.initUIVector.restype = None
+
+
+def init_uivector():
+    """
+    initUIVector: Allocate in memory an empty libscientific double vector
+    """
+    uivect = ctypes.POINTER(UIVECTOR)()
+    lsci.initUIVector(ctypes.pointer(uivect))
+    return uivect
+
+
[…]
