+++ src/python_bindings/libscientific/__init__.py
@@ -5,5 +5,6 @@ from . import tensor
 from . import pca
 from . import pls
 from . import vectlist
+from . import clustering
 
 __version__ = info.Scientific().__version__
