### 2023-04-26 #5258 @jbytecode
Dear @gmrandazzo

Thank you for submitting your manuscript and software to JOSS. I am the handling editor.

Do you want to provide me a list of potential reviewers? You can use [this list](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822).

Please perform your search from bottom to top. Please do not use the `@` character when you mention their usernames just because we don't want to bother them.

Thank you in advance.
