### 2023-04-27 #5258 @jbytecode
👋👋👋 Dear @faosorios and @mikeaalv  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**Libscientific: A Powerful C Library for Multivariate Analysis**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5258).

Thank you in advance!
