### 2023-04-27 #5258 @mikeaalv
Yes. What would be the timeline?

Thank you.
Best.
yue

On Thu, Apr 27, 2023 at 12:55 AM Mehmet Hakan Satman <
***@***.***> wrote:

> 👋👋👋 Dear @faosorios <https://github.com/faosorios> and @mikeaalv
> <https://github.com/mikeaalv> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled: "*Libscientific: A Powerful
> C Library for Multivariate Analysis*". You can find more information at
> the top of this Github issue (#5258
> <https://github.com/openjournals/joss-reviews/issues/5258>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be 23 check items for each single reviewer.
>
> Thank you in advance!
>
