### 2023-04-28 e0202f7
Adding fast matrix import from csv; Fix Tensor/Matrix input for pca/pls/cpca; Fix missing/invalid DOIs; Add badge JOSS
