+++ paper.bib
@@ -53,6 +53,7 @@
     year={2001},
     publisher={Wiley},
     address={Chichester},
+    doi={10.1088/0957-0233/12/10/708},
     isbn={0471974285; 9780471974284},
     language={eng},
     url={https://www.wiley.com/en-us/Multivariate+Analysis+of+Quality%3A+An+Introduction-p-9780471974284}
@@ -128,7 +129,7 @@ pages = {11-18},
 year = {2017},
 note = {Identification of molecules from non-targeted analysis},
 issn = {1570-0232},
-doi = {https://doi.org/10.1016/j.jchromb.2017.04.032},
+doi = {10.1016/j.jchromb.2017.04.032},
 url = {https://www.sciencedirect.com/science/article/pii/S157002321730716X},
 author = {Giuseppe Marco Randazzo and David Tonoli and Petra Strajhar and Ioannis Xenarios and Alex Odermatt and Julien Boccard and Serge Rudaz},
 keywords = {Metabolite annotation, Metabolomics, Dynamic retention time prediction, Steroidogenesis, Linear solvent strength theory, H295R cell line},
@@ -143,7 +144,7 @@ volume = {1612},
 pages = {460661},
 year = {2020},
 issn = {0021-9673},
-doi = {https://doi.org/10.1016/j.chroma.2019.460661},
+doi = {10.1016/j.chroma.2019.460661},
 url = {https://www.sciencedirect.com/science/article/pii/S0021967319310830},
 author = {Giuseppe Marco Randazzo and Andrea Bileck and Andrea Danani and Bruno Vogt and Michael Groessl},
 keywords = {3D-QSRR, Steroidomics, Deep learning, Convolutional neural network, Time-of-flight mass spectrometry, Two-dimensional gas chromatography},
