### 2023-04-28 #5420 @editorialbot
[REVIEW]: Libscientific: A Powerful C Library for Multivariate Analysis
**Submitting author:** @gmrandazzo (<a href="http://orcid.org/0000-0003-1585-0019">Giuseppe Marco Randazzo</a>)
**Repository:** https://github.com/gmrandazzo/libscientific
**Branch with paper.md** (empty if default branch):
**Version:** V1.4.1
