### 2023-04-30 gmrandazzo/libscientific#5 @mikeaalv
Difficulty in installing the package through brew
I first try the brew option.

```
% brew install --HEAD libscientific

Running `brew update --auto-update`...

^[[B==> Homebrew is run entirely by unpaid volunteers. Please consider donating:
  https://github.com/Homebrew/brew#donations

==> Downloading https://formulae.brew.sh/api/formula.jws.json
###################################################################################################################### 100.0%
==> Downloading https://formulae.brew.sh/api/cask.jws.json
###################################################################################################################### 100.0%
Warning: No available formula with the name "libscientific". Did you mean libident?
==> Searching for similarly named formulae and casks...
==> Formulae
libident

To install libident, run:
  brew install libident
```

The local installation from source seems fine. I have to modify a bit. The one-line command doesn't work for me but separating into multiple lines works. It can also be good that to mention the location the command should be run and a simple test on whether the installation is successful.
