### 2023-04-30 #5420 @mikeaalv
Better description on installation is needed (https://github.com/gmrandazzo/libscientific/issues/5)
