### 2023-04-30 gmrandazzo/libscientific#5 @jbytecode
Due to the JOSS submission: https://github.com/openjournals/joss-reviews/issues/5420
