### 2023-04-30 #5420 @mikeaalv
Hi @gmrandazzo

Thank you for providing this useful toolbox for C programmers. libscientific contains many useful tools for data analysis, particularly in chemoinformatics and metabolomics. Providing the interface to multivariate analysis in C can be helpful to build efficient pipelines. I have a few comments about the package:

1. It seems that the author starts to work on tests but no (automatic) tests are available in the repository. Even though a major part of the code is statistics-related, tests are still possible and necessary to ensure the quality for now and future. For major functions, I suggest adding tests including tests on simple toy examples and tests of invariant properties. Examples can be found in the repository of scikit-learn.

2. Coding in C doesn't remove the need of comparing performance with other tools (in C or not, e.g., python). Speed (or even memory) comparison with regular tools is needed in the manuscript.

3. I suggest adding documents for important functions in C and Python, in addition to examples in the docs and annotation in codes. Adding at least one C example (within an example folder) that the user can directly compile and run can also be helpful.

4. More background is needed. I suggest mentioning and citing previous libraries (particularly in C) that can cover similar functions or part of the functions in the Statement of need. I also suggest adding a brief description of application cases in Multivariate analysis algorithms specs, so that the reader understands the usage case.

5. Usage examples. The example "Sampling example on a drug dataset" ran after I add `time.sleep`. Having the data in other storage places might help. Both examples will benefit by having more descriptions of the chemical questions in the beginning so that the reader can understand the background.

6. Please add more details to Contributing (e.g., 2nd step). Adding a test folder to collect automatic tests can help. Adding more information related to the support and issue page can help.

7. Grammar problems. "Whether a scientists work on research or data analytics, libscientific can help gain deeper insights into the data"
