### 2023-05-08 8a492a8
Starting upgrade documentation; Adding loop unrolling for matrix-matrix dot product.
