+++ docs/Source/conf.py
@@ -53,9 +53,9 @@ copyright = u'2011, Giuseppe Marco Randazzo'
 # built documents.
 #
 # The short X.Y version.
-version = '1.2.4'
+version = '1.5.1'
 # The full version, including alpha/beta/rc tags.
-release = '1.2.4'
+release = '1.5.1'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -224,5 +224,3 @@ man_pages = [
 
 c_autodoc_roots = ['../../src']
 set_type_checking_flag = True
-
-
