+++ docs/Source/requirements.txt
@@ -1,0 +1,6 @@
+# Due to read the docs libclang linking in conf.py we want to use clang 6
+clang==14
+sphinx==5.3.0
+sphinx_rtd_theme==1.1.1
+readthedocs-sphinx-search==0.1.1
+sphinx-c-autodoc==1.1.1
