+++ docs/.readthedocs.yaml
@@ -1,0 +1,29 @@
+# .readthedocs.yaml
+# Read the Docs configuration file
+# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
+
+# Required
+version: 2
+
+# Set the version of Python and other tools you might need
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
+  apt_packages:
+    - libclang-14-dev
+
+# Build documentation in the docs/ directory with Sphinx
+sphinx:
+   configuration: docs/Source/conf.py
+   fail_on_warning: false
+
+# If using Sphinx, optionally build your docs in additional formats such as PDF
+# formats:
+#    - pdf
+formats: all
+
+# Optionally declare the Python requirements required to build your docs
+python:
+   install:
+   - requirements: docs/Source/requirements.txt
