+++ src/pls.h
@@ -25,8 +25,39 @@
 
 #define PLSCONVERGENCE 1e-8
 
-/*
+/**
  * PLS model data structure
+ * 
+ * - **xscores** x space scores
+ * - **xloadings** x space loadings
+ * - **xweights** x space weights
+ * - **yscores** y space scores
+ * - **yloadings** y space loadings
+ * - **b** pls regression coefficients
+ * - **xvarexp** variance explained in the x space
+ * - **xcolaverage** x independent variable column average
+ * - **xcolscaling** x independent variable column scaling
+ * - **ycolaverage** y independent variable column average
+ * - **ycolscaling** y independent variable column scaling
+ * - **recalculated_y** y recalculated
+ * - **recalc_residuals** y recalculated residuals
+ * - **predicted_y** y predicted 
+ * - **pred_residuals** y predicted residuals 
+ * - **r2y_recalculated** r squared using y recalculated values
+ * - **r2y_validation**
+ * - **q2y** q squared using y predicted values
+ * - **sdep** standard deviation over prediction using y predictions
+ * - **sdec** standard deviation over recalculation using y recalculated
+ * - **bias** bias
+ * - **roc_recalculated** receiver operating characteristic using y recalculated 
+ * - **roc_validation** receiver operating characteristic using y predicted 
+ * - **roc_auc_recalculated** receiver operating characteristic area under the curve using y recalculated
+ * - **roc_auc_validation** eceiver operating characteristic area under the curve using y predicted
+ * - **precision_recall_recalculated** precision-recall curve using y recalculated
+ * - **precision_recall_validation** precision-recall curve using y predicted
+ * - **precision_recall_ap_recalculated** precision-recall aread under the curve using y recalculated
+ * - **precision_recall_ap_validation** precision-recall aread under the curve using y predicted
+ * - **yscrambling** y-scrambling r-squared and q-squared
  */
 typedef struct{
   matrix *xscores;
@@ -34,7 +65,6 @@ typedef struct{
   matrix *xweights;
   matrix *yscores;
   matrix *yloadings;
-  matrix *cweights;
   dvector *b;
   dvector *xvarexp;
   dvector *xcolaverage;
@@ -52,7 +82,6 @@ typedef struct{
   matrix *sdep; /* Standard Deviation over Prediction */
   matrix *sdec; /* Standard Deviation over Recalculating */
   matrix *bias;
-
   /* Discriminant Analyisis variables */
   tensor *roc_recalculated;
   tensor *roc_validation;
@@ -62,23 +91,16 @@ typedef struct{
   tensor *precision_recall_validation;
   matrix *precision_recall_ap_recalculated;
   matrix *precision_recall_ap_validation;
-
   matrix *yscrambling;
 } PLSMODEL;
 
-/*
- * Description: Create a new PLSMODEL
- *
- * Parameters:
- *     m: PLSMODEL data structure
- *
- * Returns:
- *     The square root of 4, always.
+/**
+ * Create a new PLSMODEL
  */
 void NewPLSModel(PLSMODEL **m);
 
-/*
- * Description: Delete a PLSMODEL
+/**
+ * Delete a PLSMODEL
  */
 void DelPLSModel(PLSMODEL **m);
 
@@ -97,8 +119,34 @@ void DelPLSModel(PLSMODEL **m);
  */
 void LVCalc(matrix *X, matrix *Y, dvector *t, dvector *u, dvector *p, dvector *q, dvector *w, double *bcoef);
 
-/*
- * Description PLS calculation from P. Geladi algorithm
+
+/**
+ * @brief Calculate a partial least squares model using the NIPALS algorithm.
+ *
+ * @param [in] mx libscientific matrix data input: x independent variables
[…]
