### 2023-05-11 gmrandazzo/libscientific#5 @gmrandazzo
Thank you! Fixed with this new release.

Unfortunately, the formula has not yet been submitted to homebrew officially, so it was not working. Hence you need first to tap this third part repository.

brew tap gmrandazzo/homebrew-gmr

And after installing the software, which will now also install the python library.

brew tap gmrandazzo/homebrew-gmr

I will comment this into the JOSS submission along with all the modifications I did to fix the issues!
