### 2023-05-11 gmrandazzo/libscientific#5 @mikeaalv
It can seems I can install it by the new guideline. I assume your second command is `brew install --HEAD libscientific`
