+++ docs/Source/conf.py
@@ -13,6 +13,8 @@
 
 import sys, os
 
+sys.path.insert(0, os.path.abspath('../../src/python_bindings/'))
+
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
@@ -125,7 +127,7 @@ html_theme = 'default'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+#html_static_path = ['_static']
 
 # If not '', a 'Last updated on:' timestamp is inserted at every page bottom,
 # using the given strftime format.
@@ -221,6 +223,5 @@ man_pages = [
 ]
 
 # -- Options for sphinx_c_autodoc, the extension being documented -----------
-
 c_autodoc_roots = ['../../src']
 set_type_checking_flag = True
