+++ src/python_bindings/libscientific/clustering.py
@@ -1,21 +1,20 @@
-"""
-pls libscientific python binding
+# pls libscientific python binding
+#
+# Copyright (C) <2023>  Giuseppe Marco Randazzo
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-Copyright (C) <2023>  Giuseppe Marco Randazzo
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-"""
 import ctypes
 import os
 import libscientific.matrix as mx
@@ -29,8 +28,7 @@ lsci.MDC.argtypes = [ctypes.POINTER(mx.MATRIX),
                      ctypes.c_size_t,
                      ctypes.c_int,
                      ctypes.POINTER(vect.UIVECTOR),
-                     ctypes.c_size_t,
-                     ctypes.POINTER(ctypes.c_int)]
+                     ctypes.c_size_t]
 lsci.MDC.restype = None
 
 def most_descriptive_compound(x_input, nobjects):
@@ -43,14 +41,12 @@ def most_descriptive_compound(x_input, nobjects):
     else:
         x_input_ = x_input
 
-    ssignal = ctypes.c_int(0)
     obj_sel = vect.init_uivector()
     lsci.MDC(x_input_,
              nobjects,
              0,
              obj_sel,
-             os.cpu_count(),
-             ctypes.pointer(ssignal))
+             os.cpu_count())
     obj_sel_lst = vect.uivector_tolist(obj_sel)
     vect.del_uivector(obj_sel)
     if xalloc is True:
@@ -63,8 +59,7 @@ lsci.MaxDis_Fast.argtypes = [ctypes.POINTER(mx.MATRIX),
                              ctypes.c_size_t,
                              ctypes.c_int,
                              ctypes.POINTER(vect.UIVECTOR),
-                             ctypes.c_size_t,
-                             ctypes.POINTER(ctypes.c_int)]
+                             ctypes.c_size_t]
 lsci.MaxDis_Fast.restype = None
 
 
@@ -78,14 +73,12 @@ def max_dissimilarity_selection(x_input, nobjects):
     else:
         x_input_ = x_input
 
-    ssignal = ctypes.c_int(0)
     obj_sel = vect.init_uivector()
     lsci.MaxDis_Fast(x_input_,
                      nobjects,
                      0,
                      obj_sel,
-                     os.cpu_count(),
-                     ctypes.pointer(ssignal))
+                     os.cpu_count())
     obj_sel_lst = vect.uivector_tolist(obj_sel)
     vect.del_uivector(obj_sel)
     if xalloc is True:
[…]
