+++ src/clustering.c
@@ -82,8 +82,7 @@ void MDC(matrix* m,
          size_t n,
          int metric,
          uivector *selections,
-         size_t nthreads,
-         ssignal *s)
+         size_t nthreads)
 {
   size_t i;
   size_t j;
@@ -154,112 +153,107 @@ void MDC(matrix* m,
 
   nmdc = 0;
   while(1){
-    if(s && (*s) == SIGSCIENTIFICSTOP){
-      break;
-    }
-    else{
-      /* Find the compound with largest value in vectinfo */
+    /* Find the compound with largest value in vectinfo */
 
-      dist = vectinfo->data[0];
-      mdc = 0;
+    dist = vectinfo->data[0];
+    mdc = 0;
 
-      for(i = 0; i < vectinfo->size; i++){
-        if(vectinfo->data[i] > dist){
-          dist = vectinfo->data[i];
-          mdc = i;
-        }
-        else{
-          continue;
-        }
-      }
-
-      nmdc++;
-
-      /*now mdc is the MDC */
-
-      UIVectorAppend(selections, mdc);
-      /*Recalculate the disances of the MDC to all the other compounds 
-       * and the reciprocal ranks as aboce subtract these reciprocal 
-       * ranks from 1 and store in the rank vector R
-       */
-      /* MULTITHREAD IMPLEMENTATION */
-      size_t step = (size_t)ceil((double)m->row/(double)nthreads);
-      size_t from = 0;
-      size_t to = step;
-      for(th = 0; th < nthreads; th++){
-        args[th].from = from;
-        args[th].to = to;
-        args[th].mdc = mdc;
-        pthread_create(&threads[th], NULL, MDCWorker, (void*) &args[th]);
-
-        from = to;
-        if(from+step > m->row){
-          to = m->row;
-        }
-        else{
-          to+=step;
-        }
-      }
-
-      for(th = 0; th < nthreads; th++){
-        pthread_join(threads[th], NULL);
-      }
-
-      MatrixSort(tmprank, 0);
-
-
-      d = 2.;
-      for(i = 0; i < tmprank->row; i++){
-        j = (size_t)tmprank->data[i][1];
-        if(j == mdc){
-          rankvector->data[j] = 0.f;/* set to 0 the mdc in order to unselect this..*/
-        }
-        else{
-          rankvector->data[j] = 1 - (1 / d);
-          d += 1;
-        }
-      }
-
-      /* Multiply values in I by the corresponding values in  R. Store the result in I.*/
-      for(i = 0 ; i < vectinfo->size; i++){
-        vectinfo->data[i] *= rankvector->data[i];
-      }
-
-      /*check that all number in the I Exceded 1 if they do then go to "Find the compound with largest value in vectinfo"
-      * else stop....
-      */
-      if(n > 0){
-        if(nmdc < n){
-          continue;
-        }
-        else{
-          break;
-        }
+    for(i = 0; i < vectinfo->size; i++){
+      if(vectinfo->data[i] > dist){
+        dist = vectinfo->data[i];
+        mdc = i;
       }
       else{
-        k = 0; /* objects > 1 */
-        l = 0; /* extracted objects */
-        for(i = 0; i < vectinfo->size; i++){
-          double val = getDVectorValue(vectinfo, i);
-          if(val > 1){
-            k++;
-          }
-          else{
-            if(FLOAT_EQ(val, 0.f, EPSILON)){
-              l++;
-            }
-            else{
-              continue;
-            }
-          }
-        }
+        continue;
+      }
+    }
 
-        if(k == vectinfo->size-l){
-          continue;
+    nmdc++;
+
+    /*now mdc is the MDC */
+
+    UIVectorAppend(selections, mdc);
[…]
