+++ src/ica.c
@@ -25,16 +25,15 @@
 #include "ica.h"
 #include "vector.h"
 #include "matrix.h"
+#include "preprocessing.h"
 #include "scientificinfo.h"
 
 
 void NewICAModel(ICAMODEL **m)
 {
   (*m) = xmalloc(sizeof(ICAMODEL));
-  initMatrix(&((*m)->scores));
-  initMatrix(&((*m)->loadings));
-  initMatrix(&((*m)->dmodx));
-  initDVector(&((*m)->varexp));
+  initMatrix(&((*m)->S));
+  initMatrix(&((*m)->W));
   initDVector(&((*m)->colaverage));
   initDVector(&((*m)->colscaling));
 }
@@ -43,57 +42,76 @@ void DelICAModel(ICAMODEL** m)
 {
   DelDVector(&((*m)->colscaling));
   DelDVector(&((*m)->colaverage));
-  DelDVector(&((*m)->varexp));
-  DelMatrix(&((*m)->dmodx));
-  DelMatrix(&((*m)->loadings));
-  DelMatrix(&((*m)->scores));
+  DelMatrix(&((*m)->W));
+  DelMatrix(&((*m)->S));
   xfree((*m));
 }
 
-
-void whitening(matrix *X, matrix *X_whiten)
+void g(dvector *x)
 {
   size_t i;
-  matrix *Xcov, *D, *Dinv, *evect, *evectT, *eX, *DinveX;
-  dvector *eval;
-  NewMatrix(&Xcov, X->row, X->row);
-  //d, E = np.linalg.eigh(cov)
-  MatrixCovariance(X, Xcov);
-  initDVector(&eval);
-  initMatrix(&evect);
-  EVectEval(Xcov, eval, evect);
-  NewMatrix(&D, eval->size, eval->size);
-  for(i = 0; i < eval->size; i++)
-    D->data[i][i] = eval->data[i];
+  for(i = 0; i < x->size; i++){
+    x->data[i] =  tanh(x->data[i]);
+  }
+}
 
-  NewMatrix(&Dinv, D->row, D->col);
-  MatrixInversion(D, Dinv);
-
-  NewMatrix(&evectT, evect->col, evect->row);
-  MatrixTranspose(evect, evectT);
-
-  NewMatrix(&eX, evectT->row, X->col);
-  MatrixDotProduct(evectT, X, eX);
-
-  NewMatrix(&DinveX, Dinv->row, eX->col);
-  MatrixDotProduct(Dinv, eX, DinveX);
-
-  DelMatrix(&eX);
-  ResizeMatrix(X_whiten, evect->row, DinveX->col);
-  MatrixDotProduct(evect, DinveX, X_whiten);
-
-  DelMatrix(&DinveX);
-  DelMatrix(&evect);
-  DelMatrix(&Dinv);
-  DelMatrix(&evectT);
-  DelMatrix(&D);
-  DelDVector(&eval);
-  DelMatrix(&Xcov);
+void g_der(dvector *x)
+{
+  size_t i;
+  double v;
+  for(i = 0; i < x->size; i++){
+    v = tanh(x->data[i]);
+    x->data[i] = 1.f-(v*v);
+  }
 }
 
 void newW(dvector *w, matrix *X, dvector *new_w)
 {
+  size_t i;
+  size_t j;
+  dvector *a;
+  NewDVector(&a, X->row);
+  DVectorMatrixDotProduct(X, w, a);
+  dvector *g_a;
+  dvector *g_der_a;
+  initDVector(&g_a);
+  initDVector(&g_der_a);
+  DVectorCopy(a, g_a);
+  g(g_a);
+  DVectorCopy(a, g_der_a);
[…]
