### 2023-05-20 220ea4e
Update matrix whitening
