+++ src/tests/testica.c
@@ -52,16 +52,23 @@ void test1()
   ICAMODEL *model;
   size_t i;
   size_t j;
+
   NewMatrix(&m, 10, 3);
-  for(i = 0; i < 10; i++){
-    for(j = 0; j < 3; j++){
-      m->data[i][j] = randDouble(-2, 2);
-    }
-  }
+  m->data[0][0] = -1.; m->data[0][1] = -1.; m->data[0][2] = -2.;
+  m->data[1][0] = 2.75643348; m->data[1][1] = 3.26710563; m->data[1][2] = 4.02353911;
+  m->data[2][0] = -0.84668509; m->data[2][1] = -1.64556477; m->data[2][2] = -0.49224986;
+  m->data[3][0] = 0.52000394; m->data[3][1] = 1.92666864; m->data[3][2] = 0.44667258;
+  m->data[4][0] = -0.15235893; m->data[4][1] = -1.52062391;  m->data[4][2] = 0.32701716;
+  m->data[5][0] = 1.39949457; m->data[5][1] = 2.14419173; m->data[5][2] = 1.5436863;
+  m->data[6][0] = -2.27972909; m->data[6][1] = -2.80653121; m->data[6][2] = -3.0862603;
+  m->data[7][0] = -1.67717986; m->data[7][1] = -2.61636771; m->data[7][2] = -2.29354756;
+  m->data[8][0] = 1.2186072;  m->data[8][1] = 1.72041471; m->data[8][2] = 0.93902191;
+  m->data[9][0] = -2.28790332; m->data[9][1] = -3.14395166; m->data[9][2] = -3.43185497;
+
 
   NewICAModel(&model);
 
-  ICA(m, 1, 2, model);
+  ICA(m, 0, 2, model);
 
   PrintICA(model);
   DelICAModel(&model);

[…]
