+++ docs/.readthedocs.yaml
@@ -19,8 +19,7 @@ build:
   jobs:
     pre_install:
       - mkdir build && cd build && cmake -DCMAKE_INSTALL_PREFIX=$READTHEDOCS_VIRTUALENV_PATH .. && make install
-      - export LD_LIBRARY_PATH=$READTHEDOCS_VIRTUALENV_PATH/lib:$LD_LIBRARY_PATH
-      - echo $LD_LIBRARY_PATH
+      - echo "export LD_LIBRARY_PATH=$READTHEDOCS_VIRTUALENV_PATH/lib:$LD_LIBRARY_PATH" >> ~/.bashrc && . ~/.bashrc
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
