### 2023-05-22 8fc45a7
Modifing preprocessing to support rescaling data from colaverage/colscaling input for prediction purpouse and update ica.c. Update readthedocs and other codes for ica and matrix operations
