+++ src/tests/testmetrics.c
@@ -24,9 +24,9 @@
 #include "matrix.h"
 #include "statistic.h"
 
-void test15()
+void test13()
 {
-  puts("Test15: ManhattanDistanceCondensed");
+  puts("Test13: ManhattanDistanceCondensed");
   size_t i;
   size_t j;
   size_t indx;
@@ -72,9 +72,9 @@ void test15()
   DelMatrix(&m);
 }
 
-void test14()
+void test12()
 {
-  puts("Test14: SquaredEuclideanDistanceCondensed");
+  puts("Test12: SquaredEuclideanDistanceCondensed");
   size_t i;
   size_t j;
   size_t indx;
@@ -120,9 +120,9 @@ void test14()
   DelMatrix(&m);
 }
 
-void test13()
+void test11()
 {
-  puts("Test13: EuclideanDistanceCondensed");
+  puts("Test11: EuclideanDistanceCondensed");
   size_t i;
   size_t j;
   size_t indx;
@@ -168,9 +168,9 @@ void test13()
   DelMatrix(&m);
 }
 
-void test12()
+void test10()
 {
-  puts("Test11: ManhattanDistance");
+  puts("Test10: ManhattanDistance");
   size_t i;
   size_t j;
   matrix *m1;
@@ -217,9 +217,9 @@ void test12()
   DelMatrix(&m2);
 }
 
-void test11()
+void test9()
 {
-  puts("Test11: CosineDistance");
+  puts("Test9: CosineDistance");
   size_t i;
   size_t j;
   matrix *m1;
@@ -266,9 +266,9 @@ void test11()
   DelMatrix(&m2);
 }
 
-void test10()
+void test8()
 {
-  puts("Test10: SquaredEuclideanDistance");
+  puts("Test8: SquaredEuclideanDistance");
   size_t i;
   size_t j;
   matrix *m1;
@@ -315,9 +315,9 @@ void test10()
   DelMatrix(&m2);
 }
 
-void test9()
+void test7()
 {
-  puts("Test9: MatrixMahalanobisDistance");
+  puts("Test7: MatrixMatrixDistance");
   /* Euclidean distance between two matrix test */
   matrix *m1;
   matrix *m2;
@@ -337,16 +337,17 @@ void test9()
   m2->data[4][0] = 3; m2->data[4][1] = 3; m2->data[4][2] = 4;
   m2->data[5][0] = 5; m2->data[5][1] = 2; m2->data[5][2] = 7;
 
-  mxmxdist = MatrixMahalanobisDistance(m1, m2);
-  if(!FLOAT_EQ(mxmxdist, 1.931663, 1e-4))
+  mxmxdist = MatrixMatrixDistance(m1, m2);
+  printf("%f\n", mxmxdist);
+  if(!FLOAT_EQ(mxmxdist, 4.980518, 1e-4))
     abort();
-  puts("MatrixMahalanobisDistance: OK");
+  puts("MatrixMatrixDistance: OK");
   DelMatrix(&m1);
   DelMatrix(&m2);
 }
-void test8()
+void test6()
 {
-  puts("Test8: CosineDistance");
+  puts("Test6: CosineDistance");
   size_t i, j, maxrow, maxcol;
   matrix *m, *dist;
 
@@ -369,9 +370,9 @@ void test8()
   DelMatrix(&dist);
 }
 
-void test7()
+void test5()
 {
-  puts("Test7: SquaredEuclideanDistance");
+  puts("Test5: SquaredEuclideanDistance");
   size_t i, j, maxrow, maxcol;
   matrix *m, *dist;
 
@@ -394,9 +395,9 @@ void test7()
   DelMatrix(&dist);
 }
 
[…]
