+++ src/ica.c
@@ -67,73 +67,92 @@ void g_der(dvector *x)
     x->data[i] = 1.f-(v*v);
   }
 }
+#include<unistd.h>
 
-/* TEST THIS METHOD!! */
+/*
+ * new_w = (X_whitened * g(w^T*X_whitened)^T)/M - (g_der(w^T*X_whitened)*1_vector_size_M*w)/M
+ */
 void newW(dvector *w, matrix *X, dvector *new_w)
 {
   size_t i;
-  size_t j;
-  dvector *a;
-  NewDVector(&a, X->row);
-  MatrixDVectorDotProduct(X, w, a);
-  
-  dvector *g_a;
-  dvector *g_der_a;
-  initDVector(&g_a);
-  initDVector(&g_der_a);
-  DVectorCopy(a, g_a);
-  g(g_a);
-  DVectorCopy(a, g_der_a);
-  g_der(g_der_a);
-  double mean_g_der = 0.f;
-  for(i = 0; i < g_der_a->size; i++)
-    mean_g_der += g_der_a->data[i];
-  mean_g_der /= (double)g_der_a->size;
+  dvector *wX;
+  dvector *wX_copy;
+  dvector *XwX;
+  dvector *one;
+  dvector *g_der_one_w;
+  /* Calculation of X * g(w^T*X)^T)/M */
+  NewDVector(&wX, X->row);
+  NewDVector(&wX_copy, X->row);
+  NewDVector(&XwX, X->col);
+  NewDVector(&one, X->row);
+  DVectorSet(one, 1.f);
 
-  dvector *b;
-  NewDVector(&b, X->col);
-  
-  for(i = 0; i < X->row; i++){
-    for(j = 0; j < X->col; j++){
-      b->data[j] += X->data[i][j] * g_a->data[i];
-    }
+  /*printf("%zu x %zu x %zu x 1 = %zu x 1 \n", X->row, X->col, w->size, wX->size);*/
+  MatrixDVectorDotProduct(X, w, wX);
+  DVectorCopy(wX, wX_copy);/* copy of wX to calculate the derivative */
+  g(wX);
+  /*printf("%zu x %zu x %zu x 1 = %zu x 1 \n", X->row, X->col, wX->size, XwX->size);*/
+  DVectorMatrixDotProduct(X, wX, XwX);
+  for(i = 0; i < XwX->size; i++){
+    XwX->data[i]/=(double)X->row;
   }
 
-  /* 
-   * w_new = (X * g(np.dot(w.T, X))).mean(axis=1) - g_der(np.dot(w.T, X)).mean() * w
-   *
-   * (X * g(np.dot(w.T, X))).mean(axis=1)  = b
-   * 
-   * g_der(np.dot(w.T, X)).mean()  = mean_g_der
-   * 
+  /* Calculation of (g_der(w^T*X)*one*w)/M */
+  g_der(wX_copy);
+  double g_der_one = DVectorDVectorDotProd(wX_copy, one);
+  
+  initDVector(&g_der_one_w);
+  DVectorCopy(w, g_der_one_w);
+  
+  for(i = 0; i < w->size; i++){
+    g_der_one_w->data[i] *= g_der_one;
+    g_der_one_w->data[i] /= (double)X->row;
+  }
+  
+  /* Calculation of new_w = (X * g(w^T*X)^T)/M - (g_der(w^T*X)*one*w)/M 
+   * by using the previous calculations.
    */
-
-  for(j = 0; j < X->col; j++){
-    b->data[j] /= (double)X->row; /*.mean(axis=1)*/
-    new_w->data[j] = b->data[j] - mean_g_der * w->data[j];
-  }
-  
[…]

[…]
