### 2023-06-13 #5420 @faosorios
Hi @gmrandazzo

Your manuscript presents a C library that provides a set of routines for performing multivariate statistical analysis, with particular emphasis on methods popular in chemometrics. In addition to the library, Python bindings are available, which allow the use by a wider community of analysts. In my opinion, this kind of projects are necessary and I hope that it will be consolidated by incorporating a greater number of statistical procedures. I have reviewed the code and installed the library on my personal computer. My comments focus on the implementation of some techniques that could be improved. In what follows, I provide some comments that the author may hopefully find useful.

1. The routine for OLS (available in the files algebra.c and mlr.c) is based on forming the cross product matrices Z^t Z and Z^t Y and inverting the matrix Z^t Z by using Gauss-Jordan elimination, which is not a recommended for computations in LS problems. Recommended procedures for LS are based in the use of QR, Cholesky or SVD decompositions.
2. The routines for PCA are based on the NIPALS algorithm, possibly motivated by the popular implementation in the area of chemometrics. I am not sure why this choice was made, instead of the more traditional SVD-based option.
3. The routines for Euclidean distance as well as MSE computation can suffer from overflow/underflow. A better alternative is to use the algorithm proposed by Blue (1978) ACM Transactions on Mathematical Software 4, 15-23, or the DLASSQ routine from LAPACK.
4. The R-squared calculation in OLS is based on the formula 1 - RSS / SST, with RSS and SST denoting the sum of residual and total squares, respectively. That formula is appropriate for models with intercept which may not be the case. Another way to calculate R2 is to simply obtain the square of the correlation between the vector of responses and the vector of predicted values. It seems to me that a good option for the routine would be to check for the presence of an intercept.
5. Libscientific has LAPACK/BLAS as a requirement so it is strange from my perspective that it does not make more intensive use of the methods available in those libraries. It would be interesting to incorporate options into the code to allow the use of other calculation procedures such as those indicated in points 1, 2 and 3.
6. The above are just a few areas where implementation could be improved, I think that the manuscript could be improved by including such ideas as an opportunity for future development.
