+++ _sources/C-API.rst.txt
@@ -1,0 +1,229 @@
+***************
+C API Reference
+***************
+
+Double Vector Functions
+-----------------------
+
+.. autocstruct:: vector.h::dvector
+
+.. autocfunction:: vector.h::NewDVector
+
+.. autocfunction:: vector.h::DelDVector
+
+.. autocfunction:: vector.h::initDVector
+
+.. autocfunction:: vector.h::DVectorAppend
+
+.. autocfunction:: vector.h::DVectorRemoveAt
+
+.. autocfunction:: vector.h::PrintDVector
+
+
+Unsigned int Vector Functions
+-----------------------------
+
+.. autocstruct:: vector.h::uivector
+
+.. autocfunction:: vector.h::NewUIVector
+
+.. autocfunction:: vector.h::DelUIVector
+
+.. autocfunction:: vector.h::initUIVector
+
+.. autocfunction:: vector.h::UIVectorAppend
+
+.. autocfunction:: vector.h::UIVectorRemoveAt
+
+.. autocfunction:: vector.h::PrintUIVector
+
+
+Int Vector Functions
+--------------------
+.. autocstruct:: vector.h::ivector
+
+.. autocfunction:: vector.h::NewIVector
+
+.. autocfunction:: vector.h::DelIVector
+
+.. autocfunction:: vector.h::initIVector
+
+.. autocfunction:: vector.h::IVectorAppend
+
+.. autocfunction:: vector.h::IVectorRemoveAt
+
+.. autocfunction:: vector.h::PrintIVector
+
+
+String Vector Functions
+-----------------------
+
+.. autocstruct:: vector.h::strvector
+
+.. autocfunction:: vector.h::NewStrVector
+
+.. autocfunction:: vector.h::DelStrVector
+
+.. autocfunction:: vector.h::initStrVector
+
+.. autocfunction:: vector.h::StrVectorAppend
+
+.. autocfunction:: vector.h::PrintStrVector
+
+Matrix Functions
+----------------
+
+.. autocstruct:: matrix.h::matrix
+
+.. autocfunction:: matrix.h::NewMatrix
+
+.. autocfunction:: matrix.h::DelMatrix
+
+.. autocfunction:: matrix.h::initMatrix
+
+.. autocfunction:: matrix.h::MatrixAppendRow
+
+.. autocfunction:: matrix.h::MatrixAppendCol
+
+.. autocfunction:: matrix.h::MatrixDeleteRowAt
+
+.. autocfunction:: matrix.h::MatrixDeleteColAt
+
+.. autocfunction:: matrix.h::PrintMatrix
+
+Tensor Functions
+----------------
+
+.. autocstruct:: tensor.h::tensor
+
+.. autocfunction:: tensor.h::NewTensor
+
+.. autocfunction:: tensor.h::NewTensorMatrix
+
+.. autocfunction:: tensor.h::AddTensorMatrix
+
+.. autocfunction:: tensor.h::DelTensor
+
+.. autocfunction:: tensor.h::initTensor
+
+.. autocfunction:: tensor.h::PrintTensor
+
+PCA Functions
+-------------
+
+.. autocstruct:: pca.h::PCAMODEL
+
+.. autocfunction:: pca.h::NewPCAModel
+
+.. autocfunction:: pca.h::DelPCAModel
+
+.. autocfunction:: pca.h::PCA
+
+.. autocfunction:: pca.h::PCAScorePredictor
+
+.. autocfunction:: pca.h::PCAIndVarPredictor
+
[…]
