+++ _sources/index.rst.txt
@@ -1,0 +1,26 @@
+.. libscientific documentation master file, created by
+   sphinx-quickstart on Tue Jul 12 10:35:29 2011.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+The libscientific documentation
+===============================
+
+.. toctree::
+  :maxdepth: 2
+  
+  Overview
+  Install
+  GettingStartedInC
+  GettingStartedInPython
+  Licensing
+  C-API
+  Python-API
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
