+++ docs/conf.py
@@ -28,7 +28,8 @@ sys.path.insert(0, os.path.abspath('../src/python_bindings/'))
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['sphinx.ext.duration',
+extensions = ['sphinx.ext.githubpages',
+              'sphinx.ext.duration',
               'sphinx.ext.doctest',
               'sphinx.ext.autodoc',
               'sphinx_c_autodoc',
