+++ docs/Source/requirements.txt
@@ -1,7 +1,7 @@
 alabaster==0.7.13
 Babel==2.12.1
 beautifulsoup4==4.12.2
-certifi==2023.5.7
+certifi==2023.7.22
 charset-normalizer==3.1.0
 clang==14.0
 docutils==0.17.1
