### 2023-08-13 8b424bb
Adding unit test for python package and fixed some bugs int the python library
