+++ src/python_bindings/libscientific/loadlibrary.py
@@ -1,19 +1,20 @@
-# loadlibrary libscientific python binding
-#
-# Copyright (C) <2019>  Giuseppe Marco Randazzo
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-# 
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""loadlibrary.py method to load libscientific library
+
+Copyright (C) <2023>  Giuseppe Marco Randazzo
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
 
 import os
 import pathlib
@@ -22,27 +23,26 @@ from ctypes.util import find_library
 
 
 def get_posix_library():
+    """Get libscientific posix library location
+    """
     paths = ["/usr/lib/libscientific.so",
              "/usr/lib64/libscientific.so",
              "/usr/local/lib/libscientific.so",
-             "/usr/local/lib64/libscientific.so"]
-    
-    for p in paths:
-        if pathlib.Path(p).is_file() == True:
-            return p
-        else:
-            continue
+             "/usr/local/lib64/libscientific.so"] 
+    for path in paths:
+        if pathlib.Path(path).is_file() is True:
+            return path
     return None
 
 
 def load_libscientific_library():
     """Load the libscientific library
     """
-    LIBSCIENTIFIC_LIB_DIR = os.environ.get("LIBSCIENTIFIC_LIB_DIR")
+    lsci_lib_dir = os.environ.get("LIBSCIENTIFIC_LIB_DIR")
     lsci = None
     if os.name == "nt":
-        if LIBSCIENTIFIC_LIB_DIR:
-            lib_path = str(pathlib.Path(LIBSCIENTIFIC_LIB_DIR))+'libscientific.dll'
+        if lsci_lib_dir:
+            lib_path = str(pathlib.Path(lsci_lib_dir))+'libscientific.dll'
         else:
             lib_path = find_library("scientific")
 
@@ -58,18 +58,20 @@ def load_libscientific_library():
         else:
             libname = "libscientific.so"
 
-        if LIBSCIENTIFIC_LIB_DIR:
-            lib_path = str(pathlib.Path(LIBSCIENTIFIC_LIB_DIR))+f'/{libname}'
+        if lsci_lib_dir:
+            lib_path = str(pathlib.Path(lsci_lib_dir))+f'/{libname}'
         else:
             lib_path = get_posix_library()
             if lib_path is None:
                 lib_path = find_library("scientific")
             else:
-                message = "Please install libscientific. "
-                message += "Go to https://github.com/gmrandazzo/libscientific"
-                message += "or if it is installed please check/specify the location in LD_LIBRARY_PATH"
-                message += "or add this environment variable export LIBSCIENTIFIC_LIB_DIR=<path libscientific>"
[…]
