+++ src/python_bindings/libscientific/matrix.py
@@ -1,20 +1,20 @@
-# matrix libscientific python binding
-#
-#
-# Copyright (C) <2019>  Giuseppe Marco Randazzo
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-# 
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""matrix.py libscientific python binding
+
+Copyright (C) <2023>  Giuseppe Marco Randazzo
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
 
 import ctypes
 import csv
@@ -477,7 +477,7 @@ class Matrix():
         self.mtx = None
 
     def __getitem__(self, keys):
-        i, j = keys     
+        i, j = keys
         return self.data_ptr()[i][j]
 
     def __setitem__(self, keys, value):
@@ -586,154 +586,3 @@ class Matrix():
         Debug the matrix content
         """
         print_matrix(self.mtx)
-
-
-if __name__ in "__main__":
-    from random import random
-    from random import seed
-    seed(123)
-    print("Create a matrix of 10x2 of random values")
-    print("First prepare a list of list")
-    a = [[random() for j in range(2)] for i in range(10)]
-    print("Modify the row 4 column 1 with a string'aiuasd'")
-    a[4][1] = "aiuasd"
-
-    print("Convert the list of list matrix into matrix type")
-    m = Matrix(a)
-    m.debug()
-
-
-    print("Get value at row 1 column 1")
-    print(m[1, 1])
-    print("Set value at row 1 column 1 with value -2 ")
-    m[1, 1] = -2.
-    m.debug()
-    print("Convert the matrix into a list of list")
-    mlst = m.tolist()
-    for row in mlst:
-        print(row)
-
-    print("Append the row [99, 100]")
-    row = [99, 100]
-    m.appendrow(row)
-    m.debug()
-
-    print("Append the a col [80, 81, 82, 83, 84, 85, 86 ,87, 88, 89, 90]")
-    col = [80, 81, 82, 83, 84, 85, 86 ,87, 88, 89, 90]
-    m.appendcol(col)
-    m.debug()
-
-    print("Test matrix_dvector_dot_product")
-    d = vector.DVector([1, 2, 3])
-    r = vector.DVector([0 for i in range(11)])
-    matrix_dvector_dot_product(m.mtx, d.dvect, r.dvect)
-    r.debug()
-    del r
-    r = vector.DVector([0 for i in range(11)])
[…]

[…]
