+++ src/python_bindings/tox.ini
@@ -9,11 +9,10 @@ deps =
     numpy
 
 commands =
-    python3 -m pip install src/python_bindings/
-    coverage run --source=src/python_bindings/libscientific/ -m pytest
+    coverage run -m pytest
     coverage xml
 
 [coverage:run]
 relative_files = True
-source = src/python_bindings/libscientific/
+source = libscientific/
 branch = True
