### 2023-08-15 10dd2df
Adding more tests and removing outdated python package code
