+++ src/tests/teststatistic.c
@@ -6,30 +6,75 @@
 
 /*
  * 
- double R2(dvector *ytrue, dvector *ypred);
- double MAE(dvector *ytrue, dvector *ypred);
- double MSE(dvector *ytrue, dvector *ypred);
- double RMSE(dvector *ytrue, dvector *ypred);
- double BIAS(dvector *ytrue, dvector *ypred);$
- void Sensitivity(dvector *dtp,
-                 double thmin,
-                 double thmax,
-                 double thstep,
-                 matrix *s);
-void PositivePredictedValue(dvector *dtp,
-        dvector *dtn,
-        double thmin,
-        double thmax,
-        double thstep,
-        matrix *p);
 void MatrixCode(matrix *inmx, matrix *outmx);
 void BifactorialMatrixExpansion(matrix* inmx, matrix* outmx);
 void YatesVarEffect(matrix *mx, dvector *veff);
-
 */
+
 void test2()
 {
-  
+  puts("Test2: R2 MAE MSE RMSE Bias");
+  dvector *ytrue;
+  dvector *ypred;
+  NewDVector(&ytrue, 11);
+  NewDVector(&ypred, 11);
+
+  ytrue->data[0] = 0.812565122887111;
+  ytrue->data[1] = 0.242616766922496;
+  ytrue->data[2] = 0.229323081244868;
+  ytrue->data[3] = 0.118176252634229;
+  ytrue->data[4] = 0.168408576488025;
+  ytrue->data[5] = 0.0626776705981787;
+  ytrue->data[6] = 0.104879346780122;
+  ytrue->data[7] = 0.576538662573102;
+  ytrue->data[8] = 0.267664842692973;
+  ytrue->data[9] = 0.770655683334576;
+  ytrue->data[10] = MISSING;
+
+  ypred->data[0] = 0.905211428909207;
+  ypred->data[1] = 0.256877156553347;
+  ypred->data[2] = 0.254734909774405;
+  ypred->data[3] = 0.1558622197113;
+  ypred->data[4] = 0.240298159936659;
+  ypred->data[5] = 0.107742068728953;
+  ypred->data[6] = 0.159972261257944;
+  ypred->data[7] = 0.667423854039156;
+  ypred->data[8] = 0.350804344049622;
+  ypred->data[9] = 0.857753138117355;
+  ypred->data[10] = MISSING;
+
+  if(FLOAT_EQ(R2(ytrue, ypred), 0.9375711069105384, 1e-12)){
+    printf("R2 OK!\n");
+  }else{
+    printf("R2 ERROR!\n");
+    abort();
+  }
+
+  if(FLOAT_EQ(MSE(ytrue, ypred), 0.00438450926319716, 1e-12)){
+    printf("MSE OK!\n");
+  }else{
+    printf("MSE ERROR!\n");
+    
+    abort();
+  }
+
+  if(FLOAT_EQ(MAE(ytrue, ypred), 0.0603173534922268, 1e-12)){
+    printf("MAE OK!\n");
+  }else{
+    printf("MAE ERROR!\n");
+    
+    abort();
+  }
+
+  if(FLOAT_EQ(BIAS(ytrue, ypred), 0.0700982284901095, 1e-12)){
+    printf("BIAS OK!\n");
+  }else{
+    printf("BIAS ERROR!\n");
+    printf("%f\n", BIAS(ytrue, ypred));
+    abort();
+  }
+  DelDVector(&ytrue);
+  DelDVector(&ypred);
 }
 
 void test1()
@@ -38,6 +83,7 @@ void test1()
   /* ROC curve test */
[…]
