+++ src/tests/testnumeric.c
@@ -3,6 +3,49 @@
 #include "numeric.h"
 #include <time.h>
 
+/* MISSING TESTS
+ * void StochasticUniversalSample
+ * void RouletteWheelselection
+ * void Combinations
+ */
+void test5()
+{
+  puts("Test5: factorial of a number");
+  if(FLOAT_EQ(Factorial(12), 479001600, 1e-1)){
+    puts("Factorial OK!");
+  }
+  else{
+    puts("Factorial ERROR!");
+    abort();
+  }
+}
+
+void test4()
+{
+  puts("Test4: Random number generator");
+  srand_(12345);
+  size_t i;
+  double r[] = {3239304236.000000,
+                1907938550.0000000000,
+                3743148142.0000000000,
+                2778857055.0000000000,
+                3762851757.0000000000,
+                795925487.0000000000,
+                340349905.0000000000,
+                1721361429.0000000000,
+                2420684656.0000000000,
+                1428642710.0000000000};
+  for(i = 0; i < 10; i++){
+    double a = rand_();
+    if(FLOAT_EQ(a, r[i], 1e-2))
+      continue;
+    else{
+      printf("error random number generator rand_: %f != %f\n", a, r[i]);
+      abort();
+    }
+  }
+}
+
 void test3()
 {
   size_t i;
@@ -44,5 +87,7 @@ int main()
   test1();
   test2();
   test3();
+  test4();
+  test5();
   return 0;
 }
