+++ .github/workflows/pylint.yml
@@ -18,8 +18,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install pylint
-        cd src/python_bindings/
-        python3 setup.py install
+        pip install src/python_bindings/
     - name: Analysing the code with pylint
       run: |
         pylint src/python_bindings/libscientific/*.py
