+++ .github/workflows/pytest.yml
@@ -29,8 +29,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install pytest
         pip install numpy
-        cd src/python_bindings/
-        python3 setup.py install
+        pip install src/python_bindings/
     - name: Testing the code with pytest
       run: |
         cd src/python_bindings/
