+++ .github/workflows/pylint.yml
@@ -18,7 +18,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install pylint
-        pip install src/python_bindings/
     - name: Analysing the code with pylint
       run: |
         pylint src/python_bindings/libscientific/*.py
