### 2023-08-16 e9eebce
Improving testing for tensor.c
