+++ src/tests/testtensor.c
@@ -21,6 +21,164 @@
 #include <time.h>
 #include "tensor.h"
 #include "vector.h"
+#include "numeric.h"
+
+
+void test9()
+{
+  puts("Test 9 - Test TensorColAverage, TensorColSDEV");
+  size_t i, j;
+  tensor *t1;
+  matrix *m; 
+  matrix *r;
+  NewMatrix(&m, 3, 3);
+  m->data[0][0] = 1; m->data[0][1] = 1; m->data[0][2] = 1;
+  m->data[1][0] = 2; m->data[1][1] = 2; m->data[1][2] = 2;
+  m->data[2][0] = 3; m->data[2][1] = 3; m->data[2][2] = 3;
+
+  initTensor(&t1);
+  for(i = 0; i < 2; i++){
+    TensorAppendMatrix(t1, m);
+  }
+
+  initMatrix(&r);
+  TensorColAverage(t1, r);
+  for(i = 0; i < r->row; i++){
+    for(j = 0; j < r->col; j++){
+      if(FLOAT_EQ(r->data[i][j], 2., 1e-3)){
+        continue;;
+      }
+      else{
+        puts("TensorColAverage ERROR!");
+        abort();
+      }  
+    }
+  }
+  DelMatrix(&r);
+  initMatrix(&r);
+  TensorColSDEV(t1, r);
+  for(i = 0; i < r->row; i++){
+    for(j = 0; j < r->col; j++){
+      if(FLOAT_EQ(r->data[i][j], 1., 1e-3)){
+        continue;;
+      }
+      else{
+        puts("TensorColAverage ERROR!");
+        abort();
+      }  
+    }
+  }
+  DelMatrix(&r);
+  
+  puts("Test 9: OK.");
+  DelMatrix(&m);
+  DelTensor(&t1);
+}
+void test8()
+{
+  puts("Test 8 - Test MeanCenteredTensor");
+  size_t i, j, k;
+  tensor *t1, *tc;
+  matrix *m; 
+  NewMatrix(&m, 3, 3);
+  m->data[0][0] = 1; m->data[0][1] = 1; m->data[0][2] = 1;
+  m->data[1][0] = 2; m->data[1][1] = 2; m->data[1][2] = 2;
+  m->data[2][0] = 3; m->data[2][1] = 3; m->data[2][2] = 3;
+
+  initTensor(&t1);
+  for(i = 0; i < 2; i++){
+    TensorAppendMatrix(t1, m);
+  }
+  NewTensor(&tc, 2);
+  NewTensorMatrix(tc, 0, 3, 3);
+  NewTensorMatrix(tc, 1, 3, 3);
+  MeanCenteredTensor(t1, tc);
+  for(k = 0; k < tc->order; k++){
+    for(i = 0; i < tc->m[k]->row; i++){
+      for(j = 0; j < tc->m[k]->col; j++){
+        if(FLOAT_EQ(tc->m[k]->data[i][j], t1->m[k]->data[i][j]-2, 1e-3)){
+          continue;;
+        }
+        else{
+          puts("MeanCenteredTensor ERROR!");
+          abort();
+        }  
+      }
+    }
+  }
+  puts("Test 8: OK.");
+  DelMatrix(&m);
+  DelTensor(&tc);
+  DelTensor(&t1);
+}
+
+void test7()
+{
+  puts("Test 7 - Testing varius tensor get/set methods");
+  tensor *t1, *t2;
+  dvector *row;
+  dvector *col;
+  size_t i, j, k;
+  NewTensor(&t1, 2);
+  NewTensorMatrix(t1, 0, 1, 3);  /* i = 1 j = 3 */
+  NewTensorMatrix(t1, 1, 1, 3);
+  TensorSet(t1, 1.23456789);
+
+  NewDVector(&row, 3);
[…]
