### 2023-08-16 2269c36
Improving testing for pca.c
