### 2023-08-18 6f10e33
Adding more documentation to python code and adding loadlibrary test
