+++ src/python_bindings/libscientific/pls.py
@@ -225,27 +225,25 @@ def revert_scaling(x_input, x_avg, x_scal):
 
 def pls_beta_inference(x_input,
                        **kwargs):
-    """Predict using PLS beta coefficients and not the whole PLS model
+    """
+    Predict using PLS beta coefficients and not the whole PLS model.
 
     Parameters
     ----------
-    x_input (list):
-        input matrix to predict
-    x_averages (list):
-        column averages coming from the training x matrix
-    x_scaling (list)
-        column scaling coming from the training x matrix 
-    y_averages (list):
-        column averages coming from the training y matrix
-    y_scaling (list)
-        column scaling coming from the training y matrix 
-    beta_coeff (list):
-        pls beta coefficients extracted from the trained pls model
+    x_input : list
+        Input matrix to predict.
+    **kwargs : dict
+        Additional keyword arguments:
+        - x_averages (list): Column averages from the training x matrix.
+        - x_scaling (list): Column scaling from the training x matrix.
+        - y_averages (list): Column averages from the training y matrix.
+        - y_scaling (list): Column scaling from the training y matrix.
+        - beta_coeff (list): PLS beta coefficients extracted from the trained PLS model.
 
     Returns
     -------
-    predicted_y (list):
-        Predicted values
+    predicted_y : list
+        Predicted values.
 
     Examples
     --------
@@ -258,8 +256,6 @@ def pls_beta_inference(x_input,
                                  y_scaling=model.get_y_column_scaling(),
                                  beta_coeff=model.get_beta_coefficients(2))
     """
-
-    # Predict using beta coefficients
     x_avg = kwargs["x_averages"]
     x_scal = kwargs["x_scaling"]
     y_avg = kwargs["y_averages"]
@@ -279,28 +275,71 @@ def pls_beta_inference(x_input,
 
 class PLS():
     """
-    Partial least squares
+    Partial Least Squares (PLS) Model
+
+    This class provides methods for creating and utilizing a PLS model for regression analysis.
 
     Parameters
     ----------
-        nlv (int) : number of latent variable
-        xscaling (int) : scaling type for x matrix
-        yscaling (int) : scaling type for y matrix
-        N.B.: xscaling and yscaling can be:
-            0 -> No scaling
-            1 -> Standard deviation scaling
-            2 -> Root mean squared scaling
-            3 -> Pareto scaling
-            4 -> Range scaling
-            5 -> Level scaling
+    nlv (int) : Number of latent variables.
+    xscaling (int) : Scaling type for x matrix. Default is 1.
+    yscaling (int) : Scaling type for y matrix. Default is 0.
+
+    Note
+    ----
+    Scaling options:
+    - 0: No scaling
+    - 1: Standard deviation scaling
+    - 2: Root mean squared scaling
+    - 3: Pareto scaling
+    - 4: Range scaling
+    - 5: Level scaling
+
+    Attributes
+    ----------
+    mpls (CDataType) : The PLS model data.
+    nlv (int) : Number of latent variables.
+    xscaling (int) : Scaling type for x matrix.
+    yscaling (int) : Scaling type for y matrix.
+
+    Methods
+    -------
+    __init__(self, nlv, xscaling=1, yscaling=0)
+    __del__(self)
+    fit(self, x_input, y_input)
+    get_tscores(self)
+    get_uscores(self)
+    get_ploadings(self)
+    get_qloadings(self)
+    get_weights(self)
+    get_beta_coefficients(self, nlv : int=1)
+    get_exp_variance(self)
+    get_x_column_scaling(self)
+    get_x_averages(self)
+    get_y_column_scaling(self)
+    get_y_averages(self)
+    predict(self, x_input, nlv_=None)
[…]
