### 2023-08-18 f033f8d
Bump version to 1.5.2
