+++ performance/compile.sh
@@ -1,0 +1,4 @@
+gcc pca_tests.c  -o pca_tests -lscientific -L/usr/local/lib/ -I/usr/local/include/
+gcc pls_tests.c  -o pls_tests -lscientific -L/usr/local/lib/ -I/usr/local/include/
+gcc cpca_tests.c  -o cpca_tests -lscientific -L/usr/local/lib/ -I/usr/local/include/
+gcc mlr_tests.c -o mlr_tests -lscientific -L/usr/local/lib -I/usr/local/include
