+++ performance/make_plots.sh
@@ -1,0 +1,4 @@
+python3 plot.py pca_results.csv; mv input_vs_cputime.png pca_input_vs_cputime.png
+python3 plot.py cpca_results.csv; mv input_vs_cputime.png cpca_input_vs_cputime.png
+python3 plot.py pls_results.csv; mv input_vs_cputime.png pls_input_vs_cputime.png
+python3 plot.py mlr_results.csv; mv input_vs_cputime.png mlr_input_vs_cputime.png

[…]
