### 2023-08-18 #5420 @jbytecode
It seems the error is about png files as the error message includes `libpng error: Not a PNG file`. You can correct the error and try it on the page renderer: https://whedon.theoj.org/
