### 2023-08-18 #5420 @gmrandazzo
> It seems the error is about png files as the error message includes `libpng error: Not a PNG file`. You can correct the error and try it on the page renderer: https://whedon.theoj.org/
>
> ```
> Looks like we failed to compile the PDF with the following error: [WARNING] Could not convert image '/tmp/tex2pdf.-bcdd8186e33cd87e/d1dea6e8fa6bd5a33ea8b238363f981a98de6663.shtml': Cannot load file Jpeg Invalid marker used PNG Invalid PNG file, signature broken Bitmap Invalid Bitmap magic identifier GIF Invalid Gif signature : HDR Invalid radiance file signature Tiff Invalid endian tag value TGA Invalid bit depth (104) [WARNING] Could not convert image '/tmp/tex2pdf.-
> ```

Awesome!
