### 2023-08-18 #5420 @gmrandazzo
> Hi @gmrandazzo
>
> Thank you for providing this useful toolbox for C programmers. libscientific contains many useful tools for data analysis, particularly in chemoinformatics and metabolomics. Providing the interface to multivariate analysis in C can be helpful to build efficient pipelines. I have a few comments about the package:

Hi @mikeaalv

Thank you for your time and consideration.

>
>     1. It seems that the author starts to work on tests but no (automatic) tests are available in the repository. Even though a major part of the code is statistics-related, tests are still possible and necessary to ensure the quality for now and future. For major functions, I suggest adding tests including tests on simple toy examples and tests of invariant properties. Examples can be found in the repository of scikit-learn.

Thank you for tackling this point, which is crucial for me too. Simple and also "manually validated" toy tests are present and automatic. To perform the test, you need first to compile the source code and then run "ctest". With that command, all the tests will be automatically triggered, and every test will be checked internally to be consistent as user-defined. Moreover, additional tests have been also developed for the python package.
Since this part was not clear now, I have updated the "README" (Manual installation) and the paper itself to explain more about the unit tests technology by adding a new section named "Algorithm stability"

To give you more information, every test checks manually from a simple matrix-matrix multiplication to the final correctness of a complex algorithm like PCA/PLS/CPCA/etc. For that, I have created different tests for every method to check if the algorithm is consistent, reproducible, and correct in terms of its results. For example, matrix-matrix multiplication or matrix-vector and vice versa are checked manually with artificial tests and tested with a "large matrix" to check for memory leak and speed. PCA/PLS and so on are verified by looking at their "math" conception. For example, for PCA there is a test wich verify that given X aftear a PCA you get T (scores) and P (loadings). By multiplying T*P you should be able to fully reconstruct the original matrix X. If this is not "meet" then there is something wrong in the algorithm. Same story for PLS, consensus PCA, unfolding PCA (UPCA), etc.

>
>     2. Coding in C doesn't remove the need of comparing performance with other tools (in C or not, e.g., python). Speed (or even memory) comparison with regular tools is needed in the manuscript.

Since this library is a collection of well established and know multivariate algorithms "performance" comparison is not really necessary.
Despite that, to answer your question, I have added a new section to the paper "# Speed and Memory Comparison" that discuss that point for PCA/CPCA/PLS/MLR in wich plotting the cpu time vs the input size show a linear trend. This means tha the algorithm complexity are linear as expected and the linear behavior is also a positive sign in terms of performance, indicating that the algorithms scale well with larger input size.
Last but not least, to avoid misunderstanding, the world "performant" has been changed to "efficient".

>
>     3. I suggest adding documents for important functions in C and Python, in addition to examples in the docs and annotation in codes. Adding at least one C example (within an example folder) that the user can directly compile and run can also be helpful.

The official documentation has been updated. Examples are already present in the documentation. A detailed description of compiling and running the code is then provided. An important amount of documentation also for the python code has been also released.

>
>     4. More background is needed. I suggest mentioning and citing previous libraries (particularly in C) that can cover similar functions or part of the functions in the Statement of need. I also suggest adding a brief description of application cases in Multivariate analysis algorithms specs, so that the reader understands the usage case.

The library is specifically directed to any kind of tabular data that needs dimensionality reduction, predictive modeling, quality control and so on. To avoid misunderstanding and unwanted sectorialization of the library, I have modified the statement of need by adding the following sentence "Libscientific was designed to analyze any kind of multivariate tabular data." and the conclusion by adding the following sentence: "Incorporating Libscientific into analytical workflows may empower professionals to leverage various multivariate techniques to crack complex relationships and patterns within datasets. By offering tools for data reduction, predictive modeling, quality control, and more, as already demonstrated in previous works in -omics science and predictive modeling[@Randazzo16;@Randazzo171;@Randazzo172;@Randazzo20;@Kwon21;@Kwon22], the library can be an indispensable asset for tackling intricate challenges across various disciplines."

>
>     5. Usage examples. The example "Sampling example on a drug dataset" ran after I add `time.sleep`. Having the data in other storage places might help. Both examples will benefit by having more descriptions of the chemical questions in the beginning so that the reader can understand the background.

Thank you for the advice. However the aim of this example is only to show the usage of the library and not the chemical example itself. To address the misunderstanding I will provide new similar examples

>
>     6. Please add more details to Contributing (e.g., 2nd step). Adding a test folder to collect automatic tests can help. Adding more information related to the support and issue page can help.

Done. You can find more details on how to contribute.

>     7. Grammar problems. "Whether a scientists work on research or data analytics, libscientific can help gain deeper insights into the data"
Fixed
