### 2023-08-18 #5420 @gmrandazzo
Dear @faosorios thank you!

>Your manuscript presents a C library that provides a set of routines for performing multivariate statistical analysis, with particular emphasis on methods popular in chemometrics. In addition to the library, Python bindings are available, which allow the use by a wider community of analysts. In my opinion, this kind of projects are necessary and I hope that it will be consolidated by incorporating a greater number of statistical procedures.
I have reviewed the code and installed the library on my personal computer. My comments focus on the implementation of some techniques that could be improved. In what follows, I provide some comments that the author may hopefully find useful.

Thank you for your thoughtful comments on our manuscript. I appreciate your positive feedback regarding the C library we developed for multivariate statistical analysis, specifically chemometrics. I agree that projects like this can significantly benefit the analytical community, and I'm glad you find it necessary. I have carefully considered your specific comments and strive to address them appropriately. Your input plays a crucial role in shaping the library, and I am committed to incorporating more statistical procedures to improve and expand its scope and utility.

>    1) The routine for OLS (available in the files algebra.c and mlr.c) is based on forming the cross product matrices Z^t Z and Z^t Y and inverting the matrix Z^t Z by using Gauss-Jordan elimination, which is not a recommended for computations in LS problems.
>    Recommended procedures for LS are based in the use of QR, Cholesky or SVD decompositions.

Thank you for your insightful comment. I appreciate your considerations about using QR triangularization or SVD decomposition for the procedures under discussion.

In choosing to maintain the use of Gauss-Jordan elimination, I took into account several factors that align with the specific goals and characteristics of the task at hand:
a) The Gauss-Jordan elimination method offers the advantage of directly solving the system of linear equations, providing a solution to the least squares problem in a single step. This efficiency is particularly beneficial in specific contexts.

b) Stability is a critical aspect when dealing with numerical methods. The Gauss-Jordan elimination method offers stability, mainly when working with positively defined or diagonal dominant matrices. The reference you provided (doi:10.1145/360569.360653) supports this stability claim.

c) Handling small-pivot values is a challenge in numerical methods. My investigation revealed that the Gauss-Jordan elimination method exhibits more stability when encountering small-pivot values than QR/LU/Cholesky decompositions. This is crucial in mitigating potential numerical instability and accuracy concerns.

d) Another aspect favoring the Gauss-Jordan elimination is its natural ability to provide the inverse of a matrix. This is a valuable feature in specific scenarios.

e) Your reference to the paper (doi:10.1145/360569.360653) raises an interesting point about ill-conditioned matrices. Despite potentially yielding larger residuals, Gauss-Jordan elimination maintains a certain level of consistency compared to other methods.

Given the primary objective of obtaining solutions for a system of equations with a full square matrix, I found that while the Gauss-Jordan elimination method involves numerous row operations and can be computationally expensive for larger matrices, it aligns well with the characteristics and goals of the task. The acceptable level of error in the computed solution, as highlighted, is of similar magnitude to the solution itself.

I hope this provides further context for my decision to maintain the current implementation. I understand that different methods have their merits, and this choice best serves the specific requirements of our project.

>    2) The routines for PCA are based on the NIPALS algorithm, possibly motivated by the popular implementation in the area of chemometrics. I am not sure why this choice was made, instead of the more traditional SVD-based option.

Yes. The implementation of the PCA routine was mainly motivaded by this chemometric popular implementation. However, the other reason is that the NIPALS algorithm, compared to the SVD based one allows to compute the first "n" user decided components. Instead the SVD needs to compute all the pc.
Last but not least, in this NIPALS implementation, missing values are threated directly in the algorithm by skipping them during the computation providing thus to do calculation of PCA with missing values. This feature is available in libscientific.

>    3) The routines for Euclidean distance as well as MSE computation can suffer from overflow/underflow. A better alternative is to use the algorithm proposed by Blue (1978) ACM Transactions on Mathematical Software 4, 15-23, or the DLASSQ routine from LAPACK.

   Thank you for pointing out this out. I will address this point in the next release. I have opened an issue. Thank you!

>    4) The R-squared calculation in OLS is based on the formula 1 - RSS / SST, with RSS and SST denoting the sum of residual and total squares, respectively. That formula is appropriate for models with intercept which may not be the case. Another way to calculate R2 is to simply obtain the square of the correlation between the vector of responses and the vector of predicted values. It seems to me that a good option for the routine would be to check for the presence of an intercept.

The two different approach of calculating R2 is one in fitting and one over the predicted values. For that reason there is a score named "bias" wich try to check the presence of an intercept and how far is from the perfect solution. I would like to mantain the R2 like this and have instead another R2 method that implement your proposal. Issue is now open.

>    5) Libscientific has LAPACK/BLAS as a requirement so it is strange from my perspective that it does not make more intensive use of the methods available in those libraries. It would be interesting to incorporate options into the code to allow the use of other calculation procedures such as those indicated in points 1, 2 and 3.

 The plan is also to make an extent use of lapack library. This will be included in the future releases. Thank you!

>    The above are just a few areas where implementation could be improved, I think that the manuscript could be improved by including such ideas as an opportunity for future development.

Again, many thanks!
