### 2023-08-31 4987bd6
update paper and bibtex
