+++ paper.md
@@ -10,7 +10,7 @@ authors:
     orcid: 0000-0003-1585-0019
     affiliation: "1"
 affiliations:
- - name: independent researcher
+ - name: Independent researcher
    index: 1
 date: 17 March 2023
 bibliography: paper.bib
@@ -25,18 +25,18 @@ Libscientific is a powerful library written in C that provides a comprehensive s
 The library includes several multivariate analysis algorithms, such as principal component analysis (PCA), 
 partial least squares regression (PLS), consensus principal component analysis (CPCA), multiblock principal component analysis,
 and multiblock partial least squares (UPLS). Libscientific also includes several other tools to analyze data, such as
-cluster analysis using KMeans Hierarchical clustering and other methods to run linear algebra calculations.
-The library also provides a python foreign function to be used inside python scripts.
+cluster analysis using K-Means Hierarchical clustering and other methods to run linear algebra calculations.
+The library also provides a Python foreign function to be used inside Python scripts.
 
 
 
 # Statement of need
 
 The library is designed to be easy to use and can be integrated into any C or C++ project.
-Additionally, libscientific comes with a foreign function Python bindings, making it accessible within Python scripts and easier to perform data analysis tasks.
-One of the main advantages of libscientific is its performance and scalability. 
+Additionally, Libscientific comes with a foreign function Python bindings, making it accessible within Python scripts and easier to perform data analysis tasks.
+One of the main advantages of Libscientific is its performance and scalability. 
 This means that large data sets can be analyzed quickly and efficiently, making it an ideal choice for applications where speed is critical.
-The library depends only on lapack for SVD and eigenvalues decomposition and can be easily integrated into embedded systems.
+The library depends only on LAPACK for SVD and eigenvalues decomposition and can be easily integrated into embedded systems.
 The current library version is 1.5.2, and here is a list of the current library features:
 
 * Principal Component Analysis (PCA)
@@ -56,7 +56,7 @@ The current library version is 1.5.2, and here is a list of the current library 
 * Matrix-matrix Euclidean, Manhattan, Cosine and Mahalanobis distances
 * Numerical integration
 * Natural cubic spline interpolation and prediction
-* Linear algebra (Eigenvector/value and SVD operated by Lapack library)
+* Linear algebra (Eigenvector/value and SVD operated by LAPACK library)
 * Ordinary Least Squares solver
 * Linear equation Solver
 * Nelder-Mead Simplex Optimization
@@ -67,11 +67,11 @@ Libscientific was designed to analyze any kind of multivariate tabular data and 
 
 # Multivariate analyisis algorithms specs
 
-Principal component analysis (PCA) is one of the most commonly used methods for multivariate analysis. PCA is an unsupervised method that compresses data into low-dimensional representations that capture the dominant variation in the data. Libscientific provides a robust implementation of PCA using the NIPALS algorithm described in Geladi paper [@Geladi86]. libscientific implementation can handle data sets with many variables, few instances, and missing values.
+Principal component analysis (PCA) is one of the most commonly used methods for multivariate analysis. PCA is an unsupervised method that compresses data into low-dimensional representations that capture the dominant variation in the data. Libscientific provides a robust implementation of PCA using the NIPALS algorithm described in Geladi paper [@Geladi86]. Libscientific implementation can handle data sets with many variables, few instances, and missing values.
 
 Partial least squares (PLS) is another commonly used method for multivariate analysis. PLS is a supervised method that captures the dominant covariation between the data matrix and the target/response. Libscientific provides one version of PLS described by Geladi et al [@Geladi86]. This implementation works with single-task and multi-task regression problems.
 
[…]
