### 2023-08-31 #5420 @jbytecode
@gmrandazzo  - Here we go, we are now ready to editorial stuff.

First, I've just sent a pull request that includes minor changes in both paper manuscript and bibliograhpy.

Please review the changes and apply them if they seem to be okay.

Second, I can not see a part of comparison of the Libscientific with the other libraries. Please add a `State of the Field` or add a paragraph that compares Libscientific with the other software in somewhere else (of course, including new citations). Please be clear why somebody use or prefer Libscientific and not the others.
