### 2023-09-06 #5420 @gmrandazzo
Hi @jbytecode

PR accepted.

I have added a state of the field paragraph in the article explaining why libscientific should be preferred and not others. The main reason is that the NIPALS algorithm, in general in PLS/CPCA/PLS/and so on, allows to work with missing data without needing data imputation. Instead, if you look at the popular scikit-learn PCA or PLS, these methods use a different approach, which does not allow you to work with missing data, and you need to do a data imputation before calculation.
To justify that, I have added two more references: LittleRubin1987 and Qifa2005.

Thank you
