+++ paper.md
@@ -10,7 +10,7 @@ authors:
     orcid: 0000-0003-1585-0019
     affiliation: "1"
 affiliations:
- - name: Independent researcher
+ - name: independent researcher
    index: 1
 date: 17 March 2023
 bibliography: paper.bib
@@ -21,22 +21,20 @@ bibliography: paper.bib
 
 Multivariate analysis is a powerful technique that allows researchers to analyze and interpret data with multiple variables.
 In today's data-driven world, multivariate analysis has become essential for the exploration of complex data sets.
-Libscientific is a powerful library written in C that provides a comprehensive set of multivariate analysis tools.
+Libscientific is a powerful library written in C that provides a comprehensive set of multivariate analysis tools based on the NIPALS algorithm.
 The library includes several multivariate analysis algorithms, such as principal component analysis (PCA), 
 partial least squares regression (PLS), consensus principal component analysis (CPCA), multiblock principal component analysis,
 and multiblock partial least squares (UPLS). Libscientific also includes several other tools to analyze data, such as
-cluster analysis using K-Means Hierarchical clustering and other methods to run linear algebra calculations.
-The library also provides a Python foreign function to be used inside Python scripts.
-
-
+cluster analysis using KMeans Hierarchical clustering and other methods to run linear algebra calculations.
+The library also provides a python foreign function to be used inside python scripts.
 
 # Statement of need
 
 The library is designed to be easy to use and can be integrated into any C or C++ project.
-Additionally, Libscientific comes with a foreign function Python bindings, making it accessible within Python scripts and easier to perform data analysis tasks.
-One of the main advantages of Libscientific is its performance and scalability. 
+Additionally, libscientific comes with a foreign function Python bindings, making it accessible within Python scripts and easier to perform data analysis tasks.
+One of the main advantages of libscientific is its performance and scalability. 
 This means that large data sets can be analyzed quickly and efficiently, making it an ideal choice for applications where speed is critical.
-The library depends only on LAPACK for SVD and eigenvalues decomposition and can be easily integrated into embedded systems.
+The library depends only on lapack for SVD and eigenvalues decomposition and can be easily integrated into embedded systems.
 The current library version is 1.5.2, and here is a list of the current library features:
 
 * Principal Component Analysis (PCA)
@@ -56,7 +54,7 @@ The current library version is 1.5.2, and here is a list of the current library 
 * Matrix-matrix Euclidean, Manhattan, Cosine and Mahalanobis distances
 * Numerical integration
 * Natural cubic spline interpolation and prediction
-* Linear algebra (Eigenvector/value and SVD operated by LAPACK library)
+* Linear algebra (Eigenvector/value and SVD operated by Lapack library)
 * Ordinary Least Squares solver
 * Linear equation Solver
 * Nelder-Mead Simplex Optimization
@@ -64,14 +62,19 @@ The current library version is 1.5.2, and here is a list of the current library 
 
 Libscientific was designed to analyze any kind of multivariate tabular data and to be applied in any scientific field.
 
+# State of The field
+
+The primary objective of libscientific is to offer a library capable of performing multivariate analysis by implementing the NIPALS algorithm. This choice stems from the limitations of prevailing popular libraries like scikit-learn, which need help handling missing and noisy data effectively[@LittleRubin1987;@Qifa2005]. Notably, the NIPALS algorithm is a robust solution to these challenges, addressing issues related to data incompleteness without necessitating prior data imputation.
+Furthermore, the NIPALS algorithm conducts iterative computations of components and latent variables, leading to more efficient use of memory resources than alternative methods found in scikit-learn. Notably, it demonstrates superior computational efficiency, especially when researchers seek to analyze only a select few of the foremost principal components or latent variables.
[…]
