+++ paper.bib
@@ -59,6 +59,29 @@
     url={https://www.wiley.com/en-us/Multivariate+Analysis+of+Quality%3A+An+Introduction-p-9780471974284}
 }
 
+@Book{LittleRubin1987,
+    author={Little, Roderick J. A.
+    and Rubin, Donald B.},
+    title={Statistical Analysis with Missing Data},
+    year={1987},
+    publisher={Wiley},
+    doi={10.1002/9781119013563},
+    isbn={9781119013563; 1119013569; 9780471183860; 0471183865},
+    language={eng},
+    url={https://onlinelibrary.wiley.com/doi/book/10.1002/9781119013563}
+}
+
+@article{Qifa2005,
+  author = {Qifa Ke and Kanade, T.},
+  booktitle = {2005 IEEE Computer Society Conference on Computer Vision and Pattern Recognition (CVPR'05)}, 
+  title = {Robust L/sub 1/ norm factorization in the presence of outliers and missing data by alternative convex programming}, 
+  volume = {1},
+  number = {},
+  pages={739-746 vol. 1},
+  doi={10.1109/CVPR.2005.309},
+  year = {2005}
+}
+
 @article{Hudson96,
     author = {Hudson, Brian D. and Hyde, Richard M. and Rahr, Elizabeth and Wood, John and Osman, Julian},
     title = {Parameter Based Methods for Compound Selection from Chemical Databases},
@@ -77,7 +100,7 @@
 @article{Holliday1996,
     author={Holliday, John D.
     and Willett, Peter},
-    title={Definitions of "Dissimilarity" for Dissimilarity-Based Compound Selection},
+    title={Definitions of {\&}{\#}x201c;Dissimilarity{\&}{\#}x201d; for Dissimilarity-Based Compound Selection},
     journal={SLAS Discovery},
     year={1996},
     month={Apr},
