### 2023-09-07 #5420 @jbytecode
@gmrandazzo - Thank you for adding the new subsection now it is more clear to readers.

Please take a look at the comment and the PR:

https://github.com/gmrandazzo/libscientific/pull/10/files#r1318146613

It seems my corrections are not applied for a reason, so please apply the corrections manually:

- Please use either libscientific or Libscientific in whole of the manuscript in a consistent way.
- Please review the generated bibliography. It seems to contain some unnecessary characters or LaTeX directives. Please make the necessary changes or formatting adjustments to ensure a clean PDF output.
- All of the programming language names should be in capital form, e.g., Pyhon but not python.
- Please perform a full read and fix any typo if exists.

Thank you in advance.
