### 2023-09-07 ccc9cbc
Removing unnecessary characters due to JOSS
