+++ paper.bib
@@ -25,7 +25,7 @@
     number={1},
     pages={41-56},
     keywords={Multi-way array; Multiorder array; Principal components; PLS; Multivariate calibration},
-    abstract={Abstract The Lohm{\"o}ller?Wold decomposition of multi-way (three-way, four-way, etc.) data arrays is combined with the non-linear partial least squares (NIPALS) algorithms to provide multi-way solutions of principal components analysis (PCA) and partial least squares modelling in latent variables (PLS). The decomposition of a multi-way array is developed as the product of a score vector and a loading array, where the score vectors have the same properties as those of ordinary two-way PCA and PLS. In image analysis, the array would instead be decomposed as the product of a loading vector and an image score matrix. The resulting methods are equivalent to the method of unfolding a multi-way array to a two-way matrix followed by ordinary PCA or PLS analysis. This automatically proves the eigenvector and least squares properties of the multi-way PCA and PLS methods. The methodology is presented; the algorithms are outlined and illustrated with a small chemical example.},
+    abstract={Abstract The Lohmöller–Wold decomposition of multi-way (three-way, four-way, etc.) data arrays is combined with the non-linear partial least squares (NIPALS) algorithms to provide multi-way solutions of principal components analysis (PCA) and partial least squares modelling in latent variables (PLS). The decomposition of a multi-way array is developed as the product of a score vector and a loading array, where the score vectors have the same properties as those of ordinary two-way PCA and PLS. In image analysis, the array would instead be decomposed as the product of a loading vector and an image score matrix. The resulting methods are equivalent to the method of unfolding a multi-way array to a two-way matrix followed by ordinary PCA or PLS analysis. This automatically proves the eigenvector and least squares properties of the multi-way PCA and PLS methods. The methodology is presented; the algorithms are outlined and illustrated with a small chemical example.},
     issn={0886-9383},
     doi={10.1002/cem.1180010107},
 }
@@ -100,7 +100,7 @@
 @article{Holliday1996,
     author={Holliday, John D.
     and Willett, Peter},
-    title={Definitions of {\&}{\#}x201c;Dissimilarity{\&}{\#}x201d; for Dissimilarity-Based Compound Selection},
+    title={Definitions of "Dissimilarity" for Dissimilarity-Based Compound Selection},
     journal={SLAS Discovery},
     year={1996},
     month={Apr},
