### 2023-09-12 ee05951
Paper update due to JOSS
