+++ paper.md
@@ -1,8 +1,8 @@
 ---
-title: 'Libscientific: A Powerful C Library for Multivariate Analysis'
+title: 'libscientific: A Powerful C Library for Multivariate Analysis'
 tags:
   - C
-  - python
+  - Python
   - chemometrics
   - multivariate analysis
 authors:
@@ -21,12 +21,12 @@ bibliography: paper.bib
 
 Multivariate analysis is a powerful technique that allows researchers to analyze and interpret data with multiple variables.
 In today's data-driven world, multivariate analysis has become essential for the exploration of complex data sets.
-Libscientific is a powerful library written in C that provides a comprehensive set of multivariate analysis tools based on the NIPALS algorithm.
+libscientific is a powerful library written in C that provides a comprehensive set of multivariate analysis tools based on the NIPALS algorithm.
 The library includes several multivariate analysis algorithms, such as principal component analysis (PCA), 
 partial least squares regression (PLS), consensus principal component analysis (CPCA), multiblock principal component analysis,
-and multiblock partial least squares (UPLS). Libscientific also includes several other tools to analyze data, such as
+and multiblock partial least squares (UPLS). libscientific also includes several other tools to analyze data, such as
 cluster analysis using KMeans Hierarchical clustering and other methods to run linear algebra calculations.
-The library also provides a python foreign function to be used inside python scripts.
+The library also provides a Python foreign function to be used inside Python scripts.
 
 # Statement of need
 
@@ -60,24 +60,23 @@ The current library version is 1.5.2, and here is a list of the current library 
 * Nelder-Mead Simplex Optimization
 * Cross validation methods: Bootstrap k-fold, Leave-One-Out, Y-Scrambling
 
-Libscientific was designed to analyze any kind of multivariate tabular data and to be applied in any scientific field.
+libscientific was designed to analyze any kind of multivariate tabular data and to be applied in any scientific field.
 
 # State of The field
 
-The primary objective of libscientific is to offer a library capable of performing multivariate analysis by implementing the NIPALS algorithm. This choice stems from the limitations of prevailing popular libraries like scikit-learn, which need help handling missing and noisy data effectively[@LittleRubin1987;@Qifa2005]. Notably, the NIPALS algorithm is a robust solution to these challenges, addressing issues related to data incompleteness without necessitating prior data imputation.
+The primary objective of libscientific is to offer a library capable of performing multivariate analysis by implementing the NIPALS algorithm. This choice stems from the limitations of prevailing popular libraries like scikit-learn, which need help handling missing and noisy data effectively [@LittleRubin1987;@Qifa2005]. Notably, the NIPALS algorithm is a robust solution to these challenges, addressing issues related to data incompleteness without necessitating prior data imputation.
 Furthermore, the NIPALS algorithm conducts iterative computations of components and latent variables, leading to more efficient use of memory resources than alternative methods found in scikit-learn. Notably, it demonstrates superior computational efficiency, especially when researchers seek to analyze only a select few of the foremost principal components or latent variables.
 In summary, libscientific aims to provide a sophisticated solution for multivariate analysis, leveraging the NIPALS algorithm's strengths to surmount issues related to missing and noisy data, optimize component calculations, and enhance computational efficiency in scenarios where the analysis focuses on a limited number of critical principal components or latent variables.
 
 # Multivariate analyisis algorithms specs
 
[…]
