### 2023-09-12 #5420 @jbytecode
@gmrandazzo- It looks better now. Thank you. By the way, there are still some issues which should be fixed:

- The citation "Ke, Q., & Kanade, T. (2005). Robust l/sub 1/ norm factorization in the presence of
outliers and missing data by alternative convex programming" still have LaTeX directives. Please fix.
- There is a figure that shows computation times in pages 3-4 but this figure is not referenced from within the text. Please mention the figure in somewhere in text and set a caption and label to the figure, e.g., Figure 1. Description.
- Please perform a full read and check whether the pdf renders correctly.
- Please ping me when you've done with the manuscript.

Thank you in advance.
