+++ src/tests/testio.c
@@ -1,0 +1,701 @@
+/* testpls.c
+*
+* Copyright (C) <2016>  Giuseppe Marco Randazzo
+*
+* This program is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*
+* This program is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#include "matrix.h"
+#include "modelvalidation.h"
+#include "io.h"
+#include "pls.h"
+#include "numeric.h"
+
+void TestCPCA3(){
+    puts("Test 3 - CPCA Model Saving/Reading");
+    size_t i, j, k;
+    size_t rowsize=20;
+    uivector *colsizes;
+    tensor *t;
+    size_t npc = 5;
+    initUIVector(&colsizes);
+    UIVectorAppend(colsizes, 10);
+    UIVectorAppend(colsizes, 7);
+    UIVectorAppend(colsizes, 13);
+
+    NewTensor(&t, colsizes->size);
+    for(k = 0; k < colsizes->size; k++){
+    NewTensorMatrix(t, k, rowsize, colsizes->data[k]);
+    srand(rowsize+colsizes->data[k]);
+    for(i = 0; i < rowsize; i++){
+        for(j = 0; j < colsizes->data[k]; j++){
+        t->m[k]->data[i][j] = rand() % 100;
+        }
+    }
+    }
+
+    CPCAMODEL *model;
+    NewCPCAModel(&model);
+    CPCA(t, 1, npc, model);
+    WriteCPCA("cpca.sqlite3", model);
+
+    CPCAMODEL *mread;
+    NewCPCAModel(&mread);
+    ReadCPCA("cpca.sqlite3", mread);
+
+
+    for(k = 0; k < model->block_scores->order; k++){
+        for(j = 0; j < model->block_scores->m[k]->col; j++){
+            for(i = 0; i < model->block_scores->m[k]->row; i++){
+                if(FLOAT_EQ(model->block_scores->m[k]->data[i][j], mread->block_scores->m[k]->data[i][j], 1e-10)){
+                    continue;   
+                }
+                else{
+                    printf("%f %f\n", model->block_scores->m[k]->data[i][j], mread->block_scores->m[k]->data[i][j]);
+                    abort();
+                }
+            }
+        }
+    }
+
+    for(k = 0; k < model->block_loadings->order; k++){
+        for(j = 0; j < model->block_loadings->m[k]->col; j++){
+            for(i = 0; i < model->block_loadings->m[k]->row; i++){
+                if(FLOAT_EQ(model->block_loadings->m[k]->data[i][j], mread->block_loadings->m[k]->data[i][j], 1e-10)){
+                    continue;   
+                }
+                else{
+                    printf("%f %f\n", model->block_loadings->m[k]->data[i][j], mread->block_loadings->m[k]->data[i][j]);
+                    abort();
+                }
+            }
+        }
+    }
+
[…]
