+++ src/io.c
@@ -1,0 +1,697 @@
+#include <stdio.h>
+#include <sqlite3.h>
+
+#include "vector.h"
+#include "matrix.h"
+#include "pca.h"
+#include "pls.h"
+#include "cpca.h"
+
+static void serialize_tensor(tensor *tx, dvector *tensor_serialized)
+{
+    size_t i, j, k, c;
+    size_t tot_size = 1;
+    for(k = 0; k < tx->order; k++){
+        tot_size += 2+(tx->m[k]->row*tx->m[k]->col);        
+    }
+    DVectorResize(tensor_serialized, tot_size);
+    c = 0;
+    tensor_serialized->data[c++] = tx->order;
+    for(k = 0; k < tx->order; k++){
+        tensor_serialized->data[c++] = tx->m[k]->row;
+        tensor_serialized->data[c++] = tx->m[k]->col;
+        for(i = 0; i < tx->m[k]->row; i++){
+            for(j = 0; j < tx->m[k]->col; j++){
+                tensor_serialized->data[c++] = tx->m[k]->data[i][j];
+            }
+        }
+    }
+}
+
+static void deserialize_tensor(dvector *tensor_serialized, tensor *tx)
+{
+    size_t i, j, k, c = 0;
+    size_t order = tensor_serialized->data[c++];
+    size_t nrow, ncol;
+    for(k = 0; k < order; k++){
+        nrow = tensor_serialized->data[c++];
+        ncol = tensor_serialized->data[c++];
+        AddTensorMatrix(tx, nrow, ncol);
+        for(i = 0; i < nrow; i++){
+            for(j = 0; j < ncol; j++){
+                tx->m[k]->data[i][j] = tensor_serialized->data[c++];
+            }
+        }
+    }
+}
+
+static void serialize_matrix(matrix *mx, dvector *matrix_serialized)
+{
+    size_t i, j, c = 0;
+    DVectorResize(matrix_serialized, 2+mx->row*mx->col);
+    matrix_serialized->data[c++]= mx->row;
+    matrix_serialized->data[c++]= mx->col;
+    for(i = 0; i < mx->row; i++){
+        for(j = 0; j < mx->col; j++){
+            matrix_serialized->data[c++] = mx->data[i][j];
+        }
+    }
+}
+
+static void deserialize_matrix(dvector *matrix_serialized, matrix *mx)
+{
+    size_t i, j, c = 0;
+    size_t nrow = (size_t)matrix_serialized->data[c++];
+    size_t ncol = (size_t)matrix_serialized->data[c++]; 
+    ResizeMatrix(mx, nrow, ncol);
+    for(i = 0; i < mx->row; i++){
+        for(j = 0; j < mx->col; j++){
+            mx->data[i][j] = matrix_serialized->data[c++];
+        }
+    }
+}
+
+static void serialize_dvectorlist(dvectorlist *dvlst, dvector *dlst_serialized)
+{
+    size_t i, j, c = 0;
+    size_t tot_size = 0;
+    for(i = 0; i < dvlst->size; i++){
+        tot_size += dvlst->d[i]->size+1;
+    }
+    DVectorResize(dlst_serialized, tot_size);
+    for(i = 0; i < dvlst->size; i++){
+        dlst_serialized->data[c++] = dvlst->d[i]->size;
+        for(j = 0; j < dvlst->d[i]->size; j++){
+            dlst_serialized->data[c++] = dvlst->d[i]->data[j];
+        }
+    }
+}
+
+
+static void deserialize_dvectorlist(dvector *dlst_serialized, dvectorlist *dlst)
+{
+    size_t i, j, c = 0;
+    dvector *v;
[…]
