### 2023-09-19 b04b5ce
Adding ffi model retention function
