+++ src/python_bindings/libscientific/pls.py
@@ -272,6 +272,57 @@ def pls_beta_inference(x_input,
     return revert_scaling(pby, y_avg, y_scal)
 
 
+lsci.WritePLS.argtypes = [
+    ctypes.POINTER(ctypes.c_char),
+    ctypes.POINTER(PLSMODEL)
+]
+lsci.WritePLS.restype = None
+
+def write_pls(dbpath: str, pls: PLSMODEL):
+    """
+    write_pls(dbpath, pls):
+    
+    Writes a PLS (Partial Least Squares) model to
+    a SQLite database.
+
+    Parameters:
+        dbpath (str): The path to the SQLite database file where the PLS 
+            model will be stored.
+        pls (PLSMODEL): The libscientific data structure representing the PLS 
+            model.
+
+    Returns:
+        None
+    """
+    encoded_string = dbpath.encode('utf-8')
+    ctypes_string = ctypes.c_char_p(encoded_string)
+    lsci.WritePLS(ctypes_string, pls)
+
+
+lsci.ReadPLS.argtypes = [
+    ctypes.POINTER(ctypes.c_char),
+    ctypes.POINTER(PLSMODEL)
+]
+lsci.ReadPLS.restype = None
+
+def read_pls(dbpath: str, pls: PLSMODEL):
+    """
+    read_pls(dbpath, pls):
+    
+    Reads a PLS (Partial Least Squares) model from a SQLite database.
+
+    Parameters:
+        dbpath (str): The path to the SQLite database file from which the PLS
+            model will be read.
+        pls (PLSMODEL): The libscientific data structure where the PLS
+            model will be loaded.
+
+    Returns:
+        None
+    """
+    encoded_string = dbpath.encode('utf-8')
+    ctypes_string = ctypes.c_char_p(encoded_string)
+    lsci.ReadPLS(ctypes_string, pls)
 
 class PLS():
     """
@@ -321,7 +372,7 @@ class PLS():
     predict(self, x_input, nlv_=None)
     """
 
-    def __init__(self, nlv, xscaling=1, yscaling=0):
+    def __init__(self, nlv=2, xscaling=1, yscaling=0):
         """
         Initialize a PLS instance.
 
@@ -331,7 +382,7 @@ class PLS():
         xscaling (int) : Scaling type for x matrix. Default is 1.
         yscaling (int) : Scaling type for y matrix. Default is 0.
         """
-        self.mpls = new_pls_model()
+        self.model = new_pls_model()
         self.nlv = nlv
         self.xscaling = xscaling
         self.yscaling = yscaling
@@ -340,10 +391,10 @@ class PLS():
         """
         Clean up resources associated with the PLS instance.
         """
-        if self.mpls is not None:
-            del_pls_model(self.mpls)
-            del self.mpls
-        self.mpls = None
+        if self.model is not None:
+            del_pls_model(self.model)
+            del self.model
+        self.model = None
 
     def fit(self, x_input, y_input):
         """
@@ -374,7 +425,7 @@ class PLS():
 
         pls_algorithm(x_input_,
                       y_input_,
-                      self.mpls,
+                      self.model,
                       nlv=self.nlv,
                       x_scaling=self.xscaling,
                       y_scaling=self.yscaling)
@@ -396,7 +447,7 @@ class PLS():
         List[List[float]]
             The T-Scores.
         """
-        return mx.matrix_to_list(self.mpls.contents.xscores)
+        return mx.matrix_to_list(self.model.contents.xscores)
 
     def get_uscores(self):
         """
@@ -407,7 +458,7 @@ class PLS():
         List[List[float]]
             The U-Scores.
         """
-        return mx.matrix_to_list(self.mpls.contents.yscores)
+        return mx.matrix_to_list(self.model.contents.yscores)
 
[…]
