### 2023-09-19 98bc1a6
Adding libsqlite3 as dependence; udpdate install; updated get_posix_library() and cd/ci workflows
