+++ CMakeLists.txt
@@ -2,8 +2,8 @@ cmake_minimum_required(VERSION 3.18)
 project(libscientific LANGUAGES C)
 set(AUTHOR "Giuseppe Marco Randazzo")
 set(VERSION_MAJOR 1)
-set(VERSION_MINOR 5)
-set(VERSION_PATCH 2)
+set(VERSION_MINOR 6)
+set(VERSION_PATCH 0)
 configure_file(src/scientificconfig.h.in scientificconfig.h)
 
 if(NOT CMAKE_BUILD_TYPE)
