+++ .github/workflows/pytest.yml
@@ -18,7 +18,7 @@ jobs:
       run: |
         sudo apt-get -y update
         sudo apt-get -y upgrade
-        sudo apt-get -y install liblapack-dev cmake
+        sudo apt-get -y install liblapack-dev cmake libsqlite3-dev
         mkdir build
         cd build
         cmake .. -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=/usr/
