+++ .github/workflows/sonarqube.yml
@@ -19,7 +19,7 @@ jobs:
         run: |
           sudo apt-get -y update
           sudo apt-get -y upgrade
-          sudo apt-get -y install liblapack-dev cmake
+          sudo apt-get -y install liblapack-dev cmake libsqlite3-dev
           mkdir build
           cd build
           cmake .. -DCMAKE_EXPORT_COMPILE_COMMANDS=TRUE -DCMAKE_BUILD_TYPE=Debug -DCMAKE_INSTALL_PREFIX=/usr
