### 2023-09-19 f3524fb
Upgrade version to 1.6.0
