### 2023-09-19 be09c6d
Put version static into python bindings
