### 2023-09-30 672e042
Fix findlibrary problem when /usr/local path is used
