### 2023-10-01 90810e6
Fix memory problems io.c and tests failed
