+++ src/tests/testio.c
@@ -22,6 +22,8 @@
 #include "pls.h"
 #include "numeric.h"
 
+#define ACCEPTABILITY 1e-8
+
 void TestCPCA3(){
     puts("Test 3 - CPCA Model Saving/Reading");
     size_t i, j, k;
@@ -58,7 +60,7 @@ void TestCPCA3(){
     for(k = 0; k < model->block_scores->order; k++){
         for(j = 0; j < model->block_scores->m[k]->col; j++){
             for(i = 0; i < model->block_scores->m[k]->row; i++){
-                if(FLOAT_EQ(model->block_scores->m[k]->data[i][j], mread->block_scores->m[k]->data[i][j], 1e-10)){
+                if(FLOAT_EQ(model->block_scores->m[k]->data[i][j], mread->block_scores->m[k]->data[i][j], ACCEPTABILITY)){
                     continue;   
                 }
                 else{
@@ -72,7 +74,7 @@ void TestCPCA3(){
     for(k = 0; k < model->block_loadings->order; k++){
         for(j = 0; j < model->block_loadings->m[k]->col; j++){
             for(i = 0; i < model->block_loadings->m[k]->row; i++){
-                if(FLOAT_EQ(model->block_loadings->m[k]->data[i][j], mread->block_loadings->m[k]->data[i][j], 1e-10)){
+                if(FLOAT_EQ(model->block_loadings->m[k]->data[i][j], mread->block_loadings->m[k]->data[i][j], ACCEPTABILITY)){
                     continue;   
                 }
                 else{
@@ -85,7 +87,7 @@ void TestCPCA3(){
 
     for(j = 0; j < model->super_scores->col; j++){
         for(i = 0; i < model->super_scores->row; i++){
-            if(FLOAT_EQ(model->super_scores->data[i][j], mread->super_scores->data[i][j], 1e-10)){
+            if(FLOAT_EQ(model->super_scores->data[i][j], mread->super_scores->data[i][j], ACCEPTABILITY)){
                 continue;   
             }
             else{
@@ -97,7 +99,7 @@ void TestCPCA3(){
 
     for(j = 0; j < model->super_weights->col; j++){
         for(i = 0; i < model->super_weights->row; i++){
-            if(FLOAT_EQ(model->super_weights->data[i][j], mread->super_weights->data[i][j], 1e-10)){
+            if(FLOAT_EQ(model->super_weights->data[i][j], mread->super_weights->data[i][j], ACCEPTABILITY)){
                 continue;   
             }
             else{
@@ -110,7 +112,7 @@ void TestCPCA3(){
     
     for(j = 0; j < model->colaverage->size; j++){
         for(i = 0; i < model->colaverage->d[j]->size; i++){
-            if(FLOAT_EQ(model->colaverage->d[j]->data[i], mread->colaverage->d[j]->data[i], 1e-10)){
+            if(FLOAT_EQ(model->colaverage->d[j]->data[i], mread->colaverage->d[j]->data[i], ACCEPTABILITY)){
                 continue;   
             }
             else{
@@ -122,7 +124,7 @@ void TestCPCA3(){
 
     for(j = 0; j < model->colscaling->size; j++){
         for(i = 0; i < model->colscaling->d[j]->size; i++){
-            if(FLOAT_EQ(model->colscaling->d[j]->data[i], mread->colscaling->d[j]->data[i], 1e-10)){
+            if(FLOAT_EQ(model->colscaling->d[j]->data[i], mread->colscaling->d[j]->data[i], ACCEPTABILITY)){
                 continue;   
             }
             else{
[…]
