### 2023-10-05 f695cf9
Update python library build system to support arch and include the library itslef into the whl package
