+++ src/CMakeLists.txt
@@ -64,8 +64,7 @@ endif()
 add_library(libscientific SHARED ${Scientific_C_SRCS} ${Scientific_C_H})
 set_target_properties(libscientific PROPERTIES VERSION ${VERSION_MAJOR}.${VERSION_MINOR}.${VERSION_PATCH} SOVERSION ${VERSION_MAJOR} )
 set_target_properties(libscientific PROPERTIES OUTPUT_NAME "scientific")
-target_link_libraries(libscientific m pthread ${LAPACK_LIBRARIES} ${SQLite3_LIBRARIES})
-
+target_link_libraries(libscientific pthread ${LAPACK_LIBRARIES} ${SQLite3_LIBRARIES})
 
 if(WIN32)
   install(TARGETS libscientific ARCHIVE DESTINATION ${CMAKE_INSTALL_PREFIX}/lib LIBRARY DESTINATION ${CMAKE_INSTALL_PREFIX}/lib COMPONENT library)
@@ -76,3 +75,24 @@ else()
   install(FILES ${Scientific_C_H} DESTINATION include/scientific)
   install(FILES scientific.h DESTINATION include)
 endif()
+
+# Copy the compiled library into the python_binding library
+# in order to create a python library with a static library inside.
+set(src_dir "${CMAKE_BINARY_DIR}/src")
+set(dst_dir "${CMAKE_SOURCE_DIR}/src/python_bindings/libscientific")
+
+set(libname "libscientific")
+# Define the custom command to copy the library
+add_custom_command(
+    TARGET libscientific
+    POST_BUILD
+    COMMAND ${CMAKE_COMMAND} -E copy
+        "${src_dir}/${libname}${CMAKE_SHARED_LIBRARY_SUFFIX}"
+        "${dst_dir}/${libname}${CMAKE_SHARED_LIBRARY_SUFFIX}"
+    COMMENT "Copying ${libname} to ${dst_dir}"
+)
+
+# Optionally, create a custom target to trigger the copy as part of the build
+add_custom_target(copy_${libname}
+    DEPENDS ${libname}
+)
