### 2023-10-06 5a971ba
Update cmake to copy the library inside the python_binding package. Modified the load library to revert for the default binary library in the package. From now there will be multiple whl packages function of their architectures
