+++ paper.bib
@@ -74,7 +74,7 @@
 @article{Qifa2005,
   author = {Qifa Ke and Kanade, T.},
   booktitle = {2005 IEEE Computer Society Conference on Computer Vision and Pattern Recognition (CVPR'05)}, 
-  title = {Robust L/sub 1/ norm factorization in the presence of outliers and missing data by alternative convex programming}, 
+  title = {Robust L1 norm factorization in the presence of outliers and missing data by alternative convex programming}, 
   volume = {1},
   number = {},
   pages={739-746 vol. 1},
