+++ paper.bib
@@ -189,7 +189,7 @@ eprint = {https://www.science.org/doi/pdf/10.1126/sciadv.abf2665},
 abstract = {Combined targeting of both intra-articular infection and inflammation achieves greater chondroprotection in septic arthritis. Bacterial infections involving joints and vital organs represent a challenging clinical problem because of the two concurrent therapeutic goals of bacterial eradication and tissue preservation. In the case of septic arthritis, permanent destruction of articular cartilage by intense host inflammation is commonly seen even after successful treatment of bacterial infection. Here, we provide scientific evidence of a novel treatment modality that can protect articular cartilage and enhanced eradication of causative bacteria in septic arthritis. Locally delivered cell-penetrating antibiotics such as rifampicin effectively eradicate intracellular reservoirs of methicillin-resistant Staphylococcus aureus within joint cells. Furthermore, mitigation of intra-articular inflammation by targeting the NLRP3 (nucleotide-binding oligomerization domain-, leucine-rich repeat- and pyrin domain-containing 3) inflammasome protects articular cartilage from damage in a murine model of knee septic arthritis. Together, concurrent mitigation of intra-articular inflammation and local adjuvant targeting of intracellular bacteria represents a promising new therapeutic strategy for septic arthritis.}
 }
 
-@ARTICLE{Kwon22,
+@article{Kwon22,
 AUTHOR={Kwon, Hyuk-Kwon and Dussik, Christopher M. and Kim, Sang-Hun and Kyriakides, Themis R. and Oh, Irvin and Lee, Francis Y.},   
 TITLE={Treating ‘Septic’ With Enhanced Antibiotics and ‘Arthritis’ by Mitigation of Excessive Inflammation},      
 JOURNAL={Frontiers in Cellular and Infection Microbiology},      
@@ -202,3 +202,17 @@ ABSTRACT={Bacterial infection within the synovial joint, commonly known as septi
 }
 
 
+@article{Lorber87,
+author = {Lorber, Avraham and Wangen, Lawrence E. and Kowalski, Bruce R.},
+title = {A theoretical foundation for the PLS algorithm},
+journal = {Journal of Chemometrics},
+volume = {1},
+number = {1},
+pages = {19-31},
+keywords = {Calibration, Indirect calibration, Multivariate, Matrix decomposition, PLS, PCR},
+doi = {https://doi.org/10.1002/cem.1180010105},
+url = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/abs/10.1002/cem.1180010105},
+eprint = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/pdf/10.1002/cem.1180010105},
+abstract = {Abstract Partial least squares (PLS) modeling is an algorithm for relating one or more dependent variables to two or more independent variables. As a regression procedure it apparently evolved from the method of principal components regression (PCR) using the NIPALS algorithm, which is similar to the power method for determining the eigenvectors and eigenvalues of a matrix. This paper presents a theoretical explanation of the PLS algorithm using singular value decomposition and the power method. The relation of PLS to PCR is demonstrated, and PLS is shown to be one of a continuum of possible solutions of a similar type. These other solutions may give better prediction than either PLS or PCR under appropriate conditions.},
+year = {1987}
+}
