### 2023-10-10 #5420 @gmrandazzo
Hello @jbytecode done.

All the points were addressed. (latex directives in a citation, figure in 2 pages now is in one page, full read and ping to you now)
For your knowledge, I have added some new text in the 'algorithm stability' and 'speed and memory comparison' with a new reference to justify some sentences that were not happy to me.

Back to you.

Thanks
