### 2023-10-10 6b300df
Add PYTHONWHLALLIN feature to copy binary libraries inside the python package. This will allow the creation of a WHL which do not need to install the library iself
