+++ docs/Source/requirements.txt
@@ -1,27 +1,0 @@
-alabaster==0.7.13
-Babel==2.12.1
-beautifulsoup4==4.12.2
-certifi==2023.7.22
-charset-normalizer==3.1.0
-clang==14.0
-docutils==0.17.1
-idna==3.4
-imagesize==1.4.1
-Jinja2==3.1.2
-MarkupSafe==2.1.3
-packaging==23.1
-Pygments==2.15.1
-readthedocs-sphinx-search==0.1.1
-requests==2.31.0
-snowballstemmer==2.2.0
-soupsieve==2.4.1
-Sphinx==5.3.0
-sphinx-c-autodoc==1.1.1
-sphinx-rtd-theme==1.1.1
-sphinxcontrib-applehelp==1.0.4
-sphinxcontrib-devhelp==1.0.2
-sphinxcontrib-htmlhelp==2.0.1
-sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==1.0.3
-sphinxcontrib-serializinghtml==1.1.5
-urllib3==2.0.6
