### 2023-10-10 #5420 @jbytecode
- Line 65 : Multivariate analyisis -> analysis
- Line 98: libscientific is a numerical algorithm... -> is libscientific is an algorithm or a set of implementations of algorithms? By the way, it is not a good idea to add some piece of new text after the review and editorial stuff. I need to track and fix many things from scratch right now.
