### 2023-10-10 65578a0
Add possibility to load other libraries from local whl package path
