### 2023-10-11 fcf5443
Updated paper and bug libscientific python binding
