+++ paper.bib
@@ -210,7 +210,7 @@ volume = {1},
 number = {1},
 pages = {19-31},
 keywords = {Calibration, Indirect calibration, Multivariate, Matrix decomposition, PLS, PCR},
-doi = {https://doi.org/10.1002/cem.1180010105},
+doi = {10.1002/cem.1180010105},
 url = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/abs/10.1002/cem.1180010105},
 eprint = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/pdf/10.1002/cem.1180010105},
 abstract = {Abstract Partial least squares (PLS) modeling is an algorithm for relating one or more dependent variables to two or more independent variables. As a regression procedure it apparently evolved from the method of principal components regression (PCR) using the NIPALS algorithm, which is similar to the power method for determining the eigenvectors and eigenvalues of a matrix. This paper presents a theoretical explanation of the PLS algorithm using singular value decomposition and the power method. The relation of PLS to PCR is demonstrated, and PLS is shown to be one of a continuum of possible solutions of a similar type. These other solutions may give better prediction than either PLS or PCR under appropriate conditions.},
