### 2023-10-11 #5420 @gmrandazzo
@jbytecode many thanks! Apologies for the last-minute add. I know that is not correct. However, those sentences could have been more cool after rereading them, and I was unhappy with them. Now, they are more solid and well-justified.

Then, I have fixed everything again. This will not be an infinite loop. :)
