### 2023-10-12 #5420 @jbytecode
@gmrandazzo - Thank you, seems better now.

Could you please perform the tasks given in https://github.com/openjournals/joss-reviews/issues/5420#issuecomment-1701232046

- Create a release tag in the repo
- Use this release tag to create a Zenodo archive.
- Report the version and the Zenodo link, DOI here.

Thank you in advance.
