### 2023-10-16 #5420 @gmrandazzo
@jbytecode unfortunatelly not... https://stats.uptimerobot.com/vlYOVuWgM/ :( When I click to modify the text of the abstract i receive the following message:

The server encountered an internal error and was unable to complete your request. Either the server is overloaded or there is an error in the application.
