### 2023-10-20 #5420 @jbytecode
@gmrandazzo - The PR includes minor changes on manuscript text and citation notation [PR](https://github.com/openjournals/joss-reviews/issues/5420). Please review and merge.
