### 2023-10-20 #5420 @jbytecode
@gmrandazzo - Both the manuscript and the bibtex look good to me. I am now recommending an acceptance. Our track editor will make the final decision. Thank you for submitting your work to JOSS.

@mikeaalv, @faosorios - Also thank to you for reviewing this submission for JOSS.
