### 2023-10-25 #5420 @arfon
@gmrandazzo – I found a couple of minor typos. Could you merge this PR and I'll be able to proceed from here: https://github.com/gmrandazzo/libscientific/pull/15
